Sparse modelling with grouped and correlated features allowing for privileged information

The R package corila implements "Sparse modelling with grouped and correlated features allowing for privileged information" (Rauschenberger, 2026). This is the first implementation of a novel algorithm. It builds upon adaptive lasso regression with the glmnet-package.

Details

Use function cv.corila() for model fitting. Type library(corila) and then ?cv.corila or help("cv.corila") to open its help file.

See the vignette for further examples. Type vignette("corila") or browseVignettes("corila") to open the vignette.

This package also includes the wrapper function multiridge() for multi-penalty ridge regression with the multiridge-package.

References

Armin Rauschenberger (2026). "Sparse modelling with grouped and correlated features allowing for privileged information". In preparation.

See also

Author

Armin Rauschenberger

Examples

?cv.corila
?coef.cv.corila
?predict.cv.corila