Multi-penalty ridge regression

Fits multi-penalty ridge regression (tuning regularisation hyperparameters and estimating regression coefficients). This is a wrapper function of some functions from the multiridge-package.

Usage

multiridge(
  x,
  y,
  z,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10L,
  penalties = NULL
)

Arguments

x

predictors: \(n \times p\) matrix

y

response: \(n\)-dimensional vector

z

\(p\)-dimensional vector with entries in \(\{1, \ldots, q\}\)

family

character "linear" (or "gaussian"), "logistic" (or "binomial"), or "cox"

foldid

\(n_0\)-dimensional vector containing the fold identifiers

nfolds

positive integer specifying the number of folds (minimum \(3\), maximum \(n\))

penalties

\(q\)-dimensional vector of penalty parameters, or NULL (cross-validation)

Value

Returns an object of class "multiridge", a list with the following slots:

slots from IWLSridge() or IWLSCoxridge()

character family with value "gaussian" (also for "linear"), "binomial" (also for "logistic"), or "cox"

\(q\)-dimensional vector penalties containing optimised regularisation hyperparameters (one for each predictor group)

list indices with nfolds slots (one for each cross-validation fold), each containing the indices of the observations

list datablocks with \(q\) slots (one for each predictor group), each containing an \(n_0 \times p_k\) matrix, where \(k \in \{1, \ldots, q\}\)

\(p\)-dimensional group vector z (see argument)

list pars with slots family (see above), the \(p\)-dimensional vectors mu.x and sd.x and the scalars mu.y and sd.y

Details

The number of observations (samples) for training or testing are indicated by \(n_0\) and \(n_1\), respectively, the number of variables (features) is indicated by \(p\), and the number of variable groups is indicated by \(q\). Observations (samples) are indexed by \(i\) in \(\{1, \ldots, n\}\), variables (features) are indexed by \(j\) in \(\{1, \ldots, p\}\), and variable groups are indexed by \(k\) in \(\{1, \ldots, q\}\). The number of variables in the \(k^{\text{th}}\) group is indicated by \(p_k\), with \(\sum_{k=1}^q p_k = p\) for non-overlapping groups.

References

Mark A. van de Wiel, Mirrelijn M. van Nee and Armin Rauschenberger (2021). "Fast cross-validation for multi-penalty high-dimensional ridge regression" Journal of Computational and Graphical Statistics 30(4):835-847. doi:10.1080/10618600.2021.1904962 .

See also

Examples

# minimal example
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
multiridge(x = x, y = y, z = z)
#> $etas
#>               [,1]
#>  [1,] -0.134835510
#>  [2,]  0.067041567
#>  [3,] -0.630620386
#>  [4,] -0.810791703
#>  [5,]  0.436924947
#>  [6,]  0.140742631
#>  [7,] -0.281010940
#>  [8,] -0.206689909
#>  [9,] -0.925747551
#> [10,] -0.346786051
#> [11,] -0.274834419
#> [12,] -1.066698210
#> [13,]  0.388510527
#> [14,] -0.022074661
#> [15,] -0.452735508
#> [16,]  0.215408893
#> [17,] -0.495534006
#> [18,]  0.246622798
#> [19,] -0.240403866
#> [20,] -0.414865428
#> [21,]  0.589190379
#> [22,]  0.013186859
#> [23,]  0.665528509
#> [24,]  0.704959748
#> [25,]  0.329626678
#> [26,]  0.787953976
#> [27,]  0.377473715
#> [28,] -0.451384821
#> [29,] -0.664088780
#> [30,] -0.661964886
#> [31,] -0.278084190
#> [32,]  0.311679789
#> [33,] -0.073672453
#> [34,]  0.392320785
#> [35,]  0.513378327
#> [36,] -0.110808177
#> [37,]  0.253220578
#> [38,]  0.130070764
#> [39,]  0.293410127
#> [40,] -0.329478951
#> [41,] -0.464449263
#> [42,]  0.001672701
#> [43,]  0.306122883
#> [44,] -0.143350041
#> [45,]  0.880713824
#> [46,]  0.647404005
#> [47,]  0.733045731
#> [48,] -0.393031379
#> [49,]  0.251547907
#> [50,]  0.196182441
#> 
#> $Ypred
#>               [,1]
#>  [1,] -0.134835510
#>  [2,]  0.067041567
#>  [3,] -0.630620386
#>  [4,] -0.810791703
#>  [5,]  0.436924947
#>  [6,]  0.140742631
#>  [7,] -0.281010940
#>  [8,] -0.206689909
#>  [9,] -0.925747551
#> [10,] -0.346786051
#> [11,] -0.274834419
#> [12,] -1.066698210
#> [13,]  0.388510527
#> [14,] -0.022074661
#> [15,] -0.452735508
#> [16,]  0.215408893
#> [17,] -0.495534006
#> [18,]  0.246622798
#> [19,] -0.240403866
#> [20,] -0.414865428
#> [21,]  0.589190379
#> [22,]  0.013186859
#> [23,]  0.665528509
#> [24,]  0.704959748
#> [25,]  0.329626678
#> [26,]  0.787953976
#> [27,]  0.377473715
#> [28,] -0.451384821
#> [29,] -0.664088780
#> [30,] -0.661964886
#> [31,] -0.278084190
#> [32,]  0.311679789
#> [33,] -0.073672453
#> [34,]  0.392320785
#> [35,]  0.513378327
#> [36,] -0.110808177
#> [37,]  0.253220578
#> [38,]  0.130070764
#> [39,]  0.293410127
#> [40,] -0.329478951
#> [41,] -0.464449263
#> [42,]  0.001672701
#> [43,]  0.306122883
#> [44,] -0.143350041
#> [45,]  0.880713824
#> [46,]  0.647404005
#> [47,]  0.733045731
#> [48,] -0.393031379
#> [49,]  0.251547907
#> [50,]  0.196182441
#> 
#> $convergence
#> [1] FALSE
#> 
#> $nIt
#> [1] 1
#> 
#> $Hresl
#> $Hresl$Hmatl
#>               [,1]
#>  [1,] -0.134835510
#>  [2,]  0.067041567
#>  [3,] -0.630620386
#>  [4,] -0.810791703
#>  [5,]  0.436924947
#>  [6,]  0.140742631
#>  [7,] -0.281010940
#>  [8,] -0.206689909
#>  [9,] -0.925747551
#> [10,] -0.346786051
#> [11,] -0.274834419
#> [12,] -1.066698210
#> [13,]  0.388510527
#> [14,] -0.022074661
#> [15,] -0.452735508
#> [16,]  0.215408893
#> [17,] -0.495534006
#> [18,]  0.246622798
#> [19,] -0.240403866
#> [20,] -0.414865428
#> [21,]  0.589190379
#> [22,]  0.013186859
#> [23,]  0.665528509
#> [24,]  0.704959748
#> [25,]  0.329626678
#> [26,]  0.787953976
#> [27,]  0.377473715
#> [28,] -0.451384821
#> [29,] -0.664088780
#> [30,] -0.661964886
#> [31,] -0.278084190
#> [32,]  0.311679789
#> [33,] -0.073672453
#> [34,]  0.392320785
#> [35,]  0.513378327
#> [36,] -0.110808177
#> [37,]  0.253220578
#> [38,]  0.130070764
#> [39,]  0.293410127
#> [40,] -0.329478951
#> [41,] -0.464449263
#> [42,]  0.001672701
#> [43,]  0.306122883
#> [44,] -0.143350041
#> [45,]  0.880713824
#> [46,]  0.647404005
#> [47,]  0.733045731
#> [48,] -0.393031379
#> [49,]  0.251547907
#> [50,]  0.196182441
#> 
#> $Hresl$Mmatl
#>              [,1]
#>  [1,] -0.45087291
#>  [2,]  0.38561494
#>  [3,] -0.20063632
#>  [4,]  0.20986851
#>  [5,] -0.46261197
#>  [6,] -0.72945401
#>  [7,]  0.18715890
#>  [8,] -0.32942749
#>  [9,] -1.39507494
#> [10,]  0.50651705
#> [11,] -0.77474293
#> [12,] -0.38970990
#> [13,]  0.10512189
#> [14,]  1.16336178
#> [15,] -0.37680937
#> [16,]  1.38232765
#> [17,] -1.74855115
#> [18,] -0.87529200
#> [19,] -0.28139464
#> [20,] -0.40817680
#> [21,]  0.23627382
#> [22,] -1.14145067
#> [23,]  0.62227191
#> [24,]  1.10339227
#> [25,]  1.00458713
#> [26,] -0.10758573
#> [27,]  1.16999253
#> [28,]  0.47308086
#> [29,]  0.08309919
#> [30,] -0.69337758
#> [31,]  0.54502706
#> [32,] -0.26805815
#> [33,] -0.83683829
#> [34,] -0.97509227
#> [35,]  0.01848639
#> [36,] -0.79621856
#> [37,]  0.63857197
#> [38,]  0.51043780
#> [39,] -1.36544585
#> [40,] -0.38950198
#> [41,]  1.05639080
#> [42,]  0.46776886
#> [43,]  0.36573584
#> [44,]  1.22757696
#> [45,] -0.41072253
#> [46,]  0.13673451
#> [47,]  0.64284983
#> [48,]  0.15060435
#> [49,]  1.47576063
#> [50,] -0.46156739
#> 
#> 
#> $linearized
#>              [,1]
#>  [1,] -0.58570842
#>  [2,]  0.45265651
#>  [3,] -0.83125671
#>  [4,] -0.60092319
#>  [5,] -0.02568702
#>  [6,] -0.58871138
#>  [7,] -0.09385204
#>  [8,] -0.53611740
#>  [9,] -2.32082249
#> [10,]  0.15973100
#> [11,] -1.04957735
#> [12,] -1.45640811
#> [13,]  0.49363241
#> [14,]  1.14128711
#> [15,] -0.82954488
#> [16,]  1.59773655
#> [17,] -2.24408516
#> [18,] -0.62866920
#> [19,] -0.52179851
#> [20,] -0.82304223
#> [21,]  0.82546420
#> [22,] -1.12826381
#> [23,]  1.28780042
#> [24,]  1.80835202
#> [25,]  1.33421381
#> [26,]  0.68036825
#> [27,]  1.54746625
#> [28,]  0.02169604
#> [29,] -0.58098959
#> [30,] -1.35534246
#> [31,]  0.26694287
#> [32,]  0.04362164
#> [33,] -0.91051074
#> [34,] -0.58277148
#> [35,]  0.53186471
#> [36,] -0.90702674
#> [37,]  0.89179255
#> [38,]  0.64050856
#> [39,] -1.07203573
#> [40,] -0.71898093
#> [41,]  0.59194153
#> [42,]  0.46944156
#> [43,]  0.67185872
#> [44,]  1.08422692
#> [45,]  0.46999130
#> [46,]  0.78413852
#> [47,]  1.37589556
#> [48,] -0.24242703
#> [49,]  1.72730853
#> [50,] -0.26538494
#> 
#> $unpen
#> [1] FALSE
#> 
#> $intercept
#> [1] TRUE
#> 
#> $eta0
#> [1] 0
#> 
#> $X1
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1
#> [31,]    1
#> [32,]    1
#> [33,]    1
#> [34,]    1
#> [35,]    1
#> [36,]    1
#> [37,]    1
#> [38,]    1
#> [39,]    1
#> [40,]    1
#> [41,]    1
#> [42,]    1
#> [43,]    1
#> [44,]    1
#> [45,]    1
#> [46,]    1
#> [47,]    1
#> [48,]    1
#> [49,]    1
#> [50,]    1
#> 
#> $family
#> [1] "gaussian"
#> 
#> $penalties
#> [1] 1.197520e+06 1.711111e+01 1.368018e+06 4.364964e+01 1.112252e+06
#> 
#> $datablocks
#> $datablocks[[1]]
#>              [,1]         [,2]         [,3]        [,4]
#>  [1,]  0.05641781 -0.901539905  2.231284301  0.20334050
#>  [2,]  0.09374255  0.192029791  1.513975591 -1.81883651
#>  [3,]  1.49507408  0.312729020 -0.123135787  0.90340733
#>  [4,] -1.61947477 -0.643809363 -0.128453208 -0.00944438
#>  [5,]  0.70927230 -0.428426888  0.293488479  0.73074237
#>  [6,]  0.78591830 -0.615193292 -0.932285026 -0.55949408
#>  [7,]  1.52215751  0.483915564 -0.568898036  0.72089921
#>  [8,] -0.21869737 -1.046197024  0.154689520  0.39238514
#>  [9,] -0.57071086 -0.968619037  0.615129523  2.04006695
#> [10,] -1.19565377  0.526652901 -0.117154151  0.83847630
#> [11,] -0.63024166 -1.254700436 -0.886598198 -0.99269545
#> [12,]  0.99215164  0.702657645 -1.071598798 -0.64355554
#> [13,]  0.90728073  1.351509366 -0.213584227 -0.40439000
#> [14,] -0.36505970  0.009602677 -0.346980528  1.21337758
#> [15,]  0.14246648 -0.133243512 -0.118012008  0.27399077
#> [16,] -0.59523930  0.092127634  0.985171704 -0.07855563
#> [17,] -0.78854082 -0.367489780 -0.498241625  1.90059732
#> [18,] -0.18279980  0.306684303 -1.286420197 -0.38332352
#> [19,] -0.53003595  1.201967202 -0.147949551 -0.71688454
#> [20,] -1.10815999 -1.232605715  1.340050317  0.69643044
#> [21,]  0.56652637 -1.496088463  0.965692614  1.13191645
#> [22,] -0.45366134  0.921954810  0.000956584 -0.87714289
#> [23,] -0.53206100  0.424144839 -1.545319274  0.99176950
#> [24,] -0.21705763  1.010157281  1.578682982  1.82238486
#> [25,]  1.00345808  0.220496026  0.036213721  0.77750279
#> [26,]  0.60600116  0.350149009 -0.975745409  1.30297235
#> [27,]  0.41751143  1.116063267 -0.046791897 -0.98484456
#> [28,]  0.68519695 -1.864426990  1.794390010  1.07696576
#> [29,] -0.30714267  1.196199071  0.477969036  0.66855535
#> [30,]  1.94280439 -0.961007166 -1.094147873 -0.30855548
#> [31,] -0.80080921  0.999712517  0.777252861 -0.05131166
#> [32,]  0.84038371 -2.737536566 -1.238183655 -2.41963360
#> [33,]  2.27984978 -0.184290938  1.500450473 -0.32873763
#> [34,]  0.01637230 -0.147134997 -0.627236412 -0.17032862
#> [35,]  0.03466138  0.508483670 -1.202225122  0.56219990
#> [36,] -0.07749847  0.356580782  0.717450425  0.64341830
#> [37,]  0.76238594 -1.466024601  0.774484356 -0.79308373
#> [38,] -0.92035190 -0.364084160 -1.367786662 -0.24257800
#> [39,]  1.60947492 -0.316668461 -0.378896655 -1.77150647
#> [40,] -0.37963272  2.429090646 -0.359350193 -0.14280969
#> [41,] -0.15559317 -0.417642232  1.303864523 -1.88573849
#> [42,] -0.17517188 -1.260264511  0.041241992 -0.24529341
#> [43,] -0.04826964  1.997217463 -0.603380723  0.14038725
#> [44,]  0.12191085  0.644889162 -0.071773482  0.21672103
#> [45,] -2.98673915 -0.894496730 -1.163328953 -0.39341163
#> [46,] -1.17436176  0.374690405  1.038586656 -1.91695967
#> [47,] -1.36164244 -0.012718821 -1.648008303 -0.90169980
#> [48,]  1.51422612  0.831572064  1.905906235  0.35600657
#> [49,] -0.32087266  0.260155719 -0.085891650 -0.10514713
#> [50,] -1.38976514  0.892776753 -1.199554302 -0.45855188
#> 
#> $datablocks[[2]]
#>              [,1]        [,2]         [,3]        [,4]
#>  [1,] -0.99260548  0.64416599  0.481237594 -0.52424656
#>  [2,]  0.71096149 -0.32251240 -0.201500740  0.10883444
#>  [3,] -1.95955233  0.95944797 -1.904805624  0.45577098
#>  [4,]  0.16692958  2.44217666 -1.156901341 -1.98817700
#>  [5,] -1.43932743 -0.66986737 -0.001256525  0.25844877
#>  [6,]  1.04584276 -1.23780819  1.526969403 -0.41561234
#>  [7,] -0.63385638  1.86290029 -0.060226529  0.90896336
#>  [8,] -0.38811575  0.89582128 -0.872582802  0.26905956
#>  [9,] -2.28643221  0.36771333 -2.435966985  0.88171621
#> [10,]  0.48566427 -0.45176654 -0.295665034 -1.38337367
#> [11,]  0.14528931 -0.72393451 -0.679803053 -0.98751772
#> [12,] -0.37897964  0.33189140  0.641321034 -2.88024448
#> [13,]  0.82756627  0.16363104  0.090168828  0.86505542
#> [14,] -1.45593236  0.35259875 -0.899598638  0.77275520
#> [15,]  0.04996726 -0.84040417 -0.792671147 -0.37280794
#> [16,]  1.03419983 -1.21316250 -0.082428042  0.24769048
#> [17,]  1.42682450 -1.57874749 -0.696061083 -1.10548193
#> [18,]  0.31032332 -0.19608871  0.276199363 -0.59342863
#> [19,]  0.80884141 -0.97730370  0.213088921 -0.81544416
#> [20,] -0.67873774 -0.08532167 -1.561348243 -0.59811025
#> [21,] -0.91665713 -0.15924308  2.016605134  0.54200879
#> [22,]  0.49004557  0.66281764 -0.443373948  0.25654548
#> [23,]  1.18889606 -2.15036294  0.258800053  1.23975327
#> [24,]  1.85154147  0.91012070  0.654750134  1.44867690
#> [25,] -0.86479374 -1.21558891  0.411375461  0.48452355
#> [26,] -0.57194972  0.23587838  2.629302235 -0.98033809
#> [27,]  0.11923828 -1.02579601 -0.067447205  0.39462586
#> [28,]  2.15153689  1.86745958 -1.600916816  0.10699745
#> [29,] -1.45597512  0.39354388 -0.411429569 -0.91075995
#> [30,]  0.15719983 -0.80230173 -0.134701903 -1.51754157
#> [31,]  0.35715668  0.70753014 -0.115502460 -0.79780077
#> [32,]  1.00404526 -0.78063585 -0.515948396  0.87848690
#> [33,]  0.88236678  1.26674017  2.027723083 -1.14677202
#> [34,]  1.23952376  0.42333915  0.791219428  0.93274212
#> [35,]  0.38721796 -0.43833730  0.367125116  0.54567041
#> [36,] -0.54274049  1.50840798 -0.337049828  0.39736674
#> [37,]  0.92501625  0.87082061  1.068233703 -0.47878447
#> [38,] -1.41064265 -0.52778529  0.307159527  0.99379433
#> [39,] -0.43195695 -1.14672117 -0.680495743  1.58216017
#> [40,] -0.13004634  0.73936848 -0.506889453 -0.33132482
#> [41,]  0.14481916  0.87273020  0.478322594 -1.27705368
#> [42,]  0.52789015 -1.65640485 -1.090524569  0.04899122
#> [43,] -1.09383689 -0.70633337 -0.063060540  0.76984210
#> [44,]  0.81481800 -0.26588957 -0.623388041  0.62123870
#> [45,] -0.26308739 -0.26781133  0.956943700  1.68165040
#> [46,] -0.75078385 -0.58299119  0.513799745  1.58523326
#> [47,] -0.25667766  0.70270285  1.539245548  1.29534068
#> [48,]  0.07030911  1.10440257 -0.760062165  0.10522734
#> [49,] -1.53163160 -0.92636466  0.894214253 -0.95918297
#> [50,]  1.11028766  0.66327543  0.847801567 -0.61516708
#> 
#> $datablocks[[3]]
#>              [,1]        [,2]         [,3]         [,4]
#>  [1,] -1.14195402 -1.64101371  0.486608442  0.325816664
#>  [2,] -1.29591123 -0.69729388  0.001219334  1.376524695
#>  [3,]  0.56655281  0.73180635  1.318933418 -0.632277222
#>  [4,]  0.89361110 -0.24650701 -1.181878770  0.162431045
#>  [5,] -0.21709447 -0.13736498 -1.196770090 -0.284064919
#>  [6,] -0.44884218  0.08101794  1.885151472 -0.411605789
#>  [7,] -0.88632449 -2.35397725  0.735029905 -1.334447045
#>  [8,]  0.21042057  0.25773239  0.110364788  0.072214709
#>  [9,] -0.34850304 -0.21749273 -1.767654953  0.132856549
#> [10,]  0.36107699  0.55643115  0.325921188 -1.345737523
#> [11,] -0.14795483  0.38821333  1.127198084 -0.577403336
#> [12,] -1.06234047 -1.31444315 -1.226918457 -0.140415600
#> [13,]  0.46367875 -1.40003904 -0.392420188 -1.426628776
#> [14,] -1.89326949 -0.73596063  0.058359712 -0.619931132
#> [15,]  0.97420117 -0.56722354 -0.399985117  1.254499886
#> [16,]  2.00504217  1.25767932  0.432653567  0.112763653
#> [17,] -1.42206329  0.76349257  0.802456834  0.602671207
#> [18,] -0.74338289  0.32997162 -0.843361377 -0.870815723
#> [19,]  1.00230337 -0.94366733 -0.440773742 -1.412542529
#> [20,] -0.47445545 -0.33093815 -0.562115624  0.874458682
#> [21,] -0.90805480 -1.93326491  1.968805861 -0.984743726
#> [22,]  0.10963749  0.01224192 -0.826900339  1.134685749
#> [23,]  0.05307089  2.64153551 -0.660545478  1.603135195
#> [24,] -0.79840567  1.34072095 -0.571587102 -0.002460818
#> [25,] -1.14219278  0.22803615 -0.506729253  0.180067537
#> [26,]  0.68783806  0.74604481  0.645466963  2.168962739
#> [27,]  0.46097756 -0.78906588 -2.860961015 -0.912781930
#> [28,] -0.36646455  0.44349766 -0.760861965  0.893205471
#> [29,]  0.13267165  1.61066235  0.004336897 -1.464866342
#> [30,] -0.48770476 -0.59285857  0.297254708  1.587124974
#> [31,]  0.14497088  0.07627614 -0.179540057 -0.917626601
#> [32,] -0.62938401 -0.24454771  1.122248575  1.255705498
#> [33,]  1.13461815  0.30113261 -0.462054176 -1.306591486
#> [34,] -1.60215824 -0.43166287  0.053285832 -0.287071912
#> [35,]  2.72804451  0.56086731  1.079259186  1.295462343
#> [36,]  0.49709761 -0.19513972  1.101187524 -0.664830133
#> [37,] -0.65873206 -0.15014100 -0.339095373  1.493152399
#> [38,]  2.10672800 -1.01204718 -0.181858210  0.405017197
#> [39,] -1.19344150  1.12199148  0.045884106 -0.366739626
#> [40,]  0.44720999  0.49122421  0.694330955  0.513372321
#> [41,] -0.63625345  0.75384570  0.729457331 -1.077944129
#> [42,]  2.17986980  0.46760920  0.720467767 -1.136778142
#> [43,]  0.52481622  0.78320295 -1.354148299  1.267741461
#> [44,]  0.37362689  1.66324336  0.091478244 -0.030924977
#> [45,]  0.76201536 -1.69473013  1.903913805 -1.687142503
#> [46,] -0.91085338  1.72099919  0.011712219 -0.060371905
#> [47,]  0.59588991 -0.42088427  1.148570856  0.827150498
#> [48,]  0.35265614 -0.77488134 -1.707646107 -0.037319945
#> [49,] -0.12568196 -0.58744607 -1.098344029 -0.717406832
#> [50,] -0.22720303  0.08311493  0.620592147  1.172450126
#> 
#> $datablocks[[4]]
#>              [,1]        [,2]        [,3]        [,4]
#>  [1,]  0.07245538  0.21272730 -0.22257381 -0.54061925
#>  [2,] -0.25328030  0.13527107 -0.97099026 -0.02484327
#>  [3,]  0.48243188 -0.05086617  0.67306545 -0.58300382
#>  [4,]  0.54147726 -0.49835321 -0.79481056  0.99703490
#>  [5,] -0.54029468 -2.31780515 -1.97636564 -1.60307882
#>  [6,] -1.73192652  1.29021409 -0.38141908 -1.26620268
#>  [7,]  0.28051839  1.49642747  2.24948127  0.01661620
#>  [8,]  2.13545733 -0.08669032 -0.41420024 -1.13279757
#>  [9,] -2.10020301  1.85700221  0.77802735  0.93146339
#> [10,]  1.32253071  0.44303052 -1.10920256  0.48984809
#> [11,]  1.28754429 -0.02065028 -1.44934362  0.09619307
#> [12,] -1.07336561  2.32553265  0.60134631  0.57077736
#> [13,]  0.91333616 -0.22035639 -0.48041951 -0.15839770
#> [14,] -0.49560465 -0.71329804  0.83785114  1.16030407
#> [15,]  0.62286880 -0.16958780  0.39476129 -2.77509090
#> [16,] -0.04289975 -0.54536002 -0.75459629 -1.30658055
#> [17,]  0.76829531  1.04319726  0.07860785  0.29095458
#> [18,] -0.12057121 -1.05130147 -0.10354371  2.30322716
#> [19,]  0.39119565  1.09005668  0.27524657  0.29576017
#> [20,] -0.18181579 -1.32610911 -0.34737433  0.18508689
#> [21,] -1.73870418  1.67796629 -0.90014189  1.82378488
#> [22,]  0.89013866 -1.05374063  1.01644298 -0.38747170
#> [23,] -0.06431416 -0.48905998  0.75105705  0.85975866
#> [24,] -1.09844029 -0.76354574  0.72945982  0.63462263
#> [25,]  2.12528530 -0.21669888  0.70583403  0.06980758
#> [26,] -0.72789668 -1.46905568 -0.54872242  1.46647316
#> [27,] -0.57638450 -0.50450946 -1.47894326  0.76738678
#> [28,]  0.24400168 -0.15182499 -0.51405305 -1.26870542
#> [29,] -0.05802546  0.59965311  1.88695781  0.65374919
#> [30,]  0.64111916  1.23047217  0.88638943 -0.01821983
#> [31,] -0.53881037 -0.62475785  0.63428245 -0.39395953
#> [32,]  0.44033810 -1.62718397  1.62617947 -0.14676803
#> [33,]  0.47482537  2.17256697 -0.75307547 -0.77483594
#> [34,] -1.56833644  0.47805204  0.93225336  0.87095415
#> [35,] -0.53150260 -0.88238187 -0.31498335  1.04434920
#> [36,] -1.31318836  0.23728569 -0.45129115  0.28690310
#> [37,] -1.42687195  0.47030899 -1.76105821  0.95417072
#> [38,] -0.21056268  0.47764819  1.75173140  0.10583635
#> [39,] -0.33561014 -0.22754548 -0.55933287 -1.11432348
#> [40,]  1.74192916  0.23726980  0.99730027  0.33791129
#> [41,] -0.99252577  0.35486367  0.23518461 -1.54088625
#> [42,] -0.13523655 -0.09920841 -1.57236910  0.94130783
#> [43,]  1.35497004 -0.28946785 -0.36727929  0.04457385
#> [44,]  0.31479415  0.85893717  0.43414971 -0.52160671
#> [45,]  0.33422663 -0.54913865 -1.32527414 -1.22723981
#> [46,] -1.73471952 -0.52549819  0.42803581  0.80265795
#> [47,]  0.74086661  0.15162225 -0.19463792 -0.78554368
#> [48,]  1.11628096  0.33508459  0.48912551 -1.39779866
#> [49,]  0.32598612 -0.83930644 -1.15386431  0.86088931
#> [50,]  0.02821806 -1.86188817  1.50709509 -0.89442890
#> 
#> $datablocks[[5]]
#>               [,1]        [,2]         [,3]        [,4]
#>  [1,]  0.977587151 -0.87468897  1.452030306 -1.93789850
#>  [2,]  0.889085525 -1.34998423  0.273028615 -1.03302631
#>  [3,] -0.870864727  3.29729195 -0.711659843  1.04433940
#>  [4,] -1.280425867 -0.14399405 -0.031278336  0.82274774
#>  [5,]  0.274619012  1.09799940  1.725054478  1.14754905
#>  [6,] -0.809444636 -0.06220726  2.250108964  1.22297597
#>  [7,]  0.073172015  0.91013507 -0.635581838 -0.21213666
#>  [8,]  0.318559038 -0.83336528  1.696506175  1.50559926
#>  [9,]  0.758299991  0.26343481  0.046907576  2.47805878
#> [10,] -0.372906298 -0.10049100 -0.782311537 -1.43323046
#> [11,]  0.090857090 -1.14687152 -0.541532345 -0.18878001
#> [12,] -1.554967893 -0.94241509 -0.140376617 -1.11777645
#> [13,] -0.082085927 -0.54095545 -2.637794674  0.26640944
#> [14,]  0.360046048 -0.03780810  0.578671758 -2.20247350
#> [15,]  1.976733162  0.06364666 -0.410471256 -2.06392162
#> [16,] -0.125229215 -0.11936950 -0.831830134  0.36385967
#> [17,] -0.472448427  0.62439941 -0.515304160  0.12893279
#> [18,] -1.036657085 -0.44206299 -0.484298184 -0.06860796
#> [19,] -0.908425139 -0.51959157 -1.227917179 -1.26577033
#> [20,] -0.243685691 -0.13775691 -0.163634631  0.49072118
#> [21,]  0.145033686 -0.26618022  2.357040081  0.75411209
#> [22,] -0.467504044  1.28410702 -1.729450447  1.22486992
#> [23,] -0.131631281  1.22457722  0.273882091  0.50083063
#> [24,]  0.824681277 -0.63036621  1.257001330  0.59776789
#> [25,]  1.260643433  1.81515834 -0.668325626 -1.14978780
#> [26,]  1.135973526  1.58494323 -1.040420834  0.97620441
#> [27,]  0.044136543 -0.01224494  1.451351737  0.27894277
#> [28,]  1.035264652 -1.53802096 -0.542773234 -0.85978838
#> [29,]  0.691953352 -0.89256487  0.387628530  0.36325485
#> [30,] -1.643854809  0.04495426  0.062373939  0.36740810
#> [31,] -0.850430049 -0.37701016 -1.486185890 -0.24733642
#> [32,] -0.293048119 -1.18909891  0.627698144 -0.79847136
#> [33,]  0.008101859  1.20638538 -0.405246058 -0.61126673
#> [34,] -0.261602084  0.75968815 -0.340024163  0.27457472
#> [35,] -1.112146855 -1.28667722  0.479083914 -0.16101569
#> [36,]  0.236006048 -0.37388817  0.760229965 -1.04878766
#> [37,] -0.903004055 -0.09556204  0.607771088  1.24676688
#> [38,] -0.211907814  2.10222591 -0.005962811  0.89313234
#> [39,]  1.345307674 -0.50604193 -1.264551749  0.04689971
#> [40,]  0.745468799  0.35269531  0.095132913  0.09536577
#> [41,] -0.830238221  0.79479052  0.439644794 -0.45241326
#> [42,] -0.282581131 -0.22678554 -0.833422277 -1.08831914
#> [43,]  0.674268382 -0.11793652  0.363626163  1.09237082
#> [44,] -1.131083180 -0.62609821 -1.078314662 -0.68196986
#> [45,]  1.373877331 -0.85803252  0.051868500 -0.49563264
#> [46,] -3.338592198 -1.48179745  0.323860211 -0.34003860
#> [47,]  0.094240322 -0.21136635 -0.182879325  0.57630639
#> [48,]  1.136220229  0.93727761  0.316133788  1.37673959
#> [49,]  0.840591836 -1.07828042 -0.355484240 -0.54847008
#> [50,]  1.904036765  0.65580432  1.170396991 -0.12982075
#> 
#> 
#> $indices
#> $indices[[1]]
#> [1]  4 13 21 46 50
#> 
#> $indices[[2]]
#> [1]  5  7 18 20 26
#> 
#> $indices[[3]]
#> [1] 10 28 32 33 44
#> 
#> $indices[[4]]
#> [1] 12 17 27 31 34
#> 
#> $indices[[5]]
#> [1]  8 16 29 38 42
#> 
#> $indices[[6]]
#> [1]  1  3  9 23 24
#> 
#> $indices[[7]]
#> [1] 25 30 37 40 41
#> 
#> $indices[[8]]
#> [1]  6 11 14 45 48
#> 
#> $indices[[9]]
#> [1]  2 15 22 35 49
#> 
#> $indices[[10]]
#> [1] 19 36 39 43 47
#> 
#> 
#> $z
#>  [1] 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5
#> 
#> $pars
#> $pars$family
#> [1] "gaussian"
#> 
#> $pars$mu.x
#>  [1] -0.083057365  0.007140121  0.110379803  0.025533828  0.184313248
#>  [6] -0.427576164  0.068349798 -0.059593388  0.195749669  0.010458997
#> [11]  0.212557138 -0.032698322  0.083237742  0.049167157  0.253638978
#> [16] -0.246516821  0.062602377 -0.030437331  0.047120976 -0.050161606
#> 
#> $pars$sd.x
#>  [1] 0.9970835 1.0143226 0.8938434 1.0557786 1.1488296 1.0589903 0.9793481
#>  [8] 1.1531980 0.8925094 1.0487883 0.9478179 0.9310024 0.9943137 1.1382961
#> [15] 1.0432079 0.9201300 1.0521665 0.8032634 0.9912460 0.8637782
#> 
#> $pars$mu.y
#> [1] 0.0105799
#> 
#> $pars$sd.y
#> [1] 1.042445
#> 
#> 
#> attr(,"class")
#> [1] "multiridge"

# \donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x %*% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
#>     glmnet multiridge 
#>  0.4692074  0.5982584 
sapply(coef, function(x) mean((beta-x[-1])^2))
#>     glmnet multiridge 
#>  0.1718059  0.1262343 

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
#>     glmnet multiridge 
#>   73.65774   50.10826 
# }