Multi-penalty ridge regression

Fits multi-penalty ridge regression (tuning regularisation hyperparameters and estimating regression coefficients). This is a wrapper function of some functions from the multiridge-package.

Usage

multiridge(
  x,
  y,
  z,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10L,
  penalties = NULL
)

Arguments

x

predictors: \(n \times p\) matrix

y

response: \(n\)-dimensional vector

z

\(p\)-dimensional vector with entries in \(\{1, \ldots, q\}\)

family

character "linear" (or "gaussian"), "logistic" (or "binomial"), or "cox"

foldid

\(n_0\)-dimensional vector containing the fold identifiers (minimum \(1\), maximum nfolds)

nfolds

positive integer specifying the number of folds (minimum \(3\), maximum \(n\))

penalties

\(q\)-dimensional vector of penalty parameters, or NULL (cross-validation)

Value

Returns an object of class "multiridge", a list with the following slots:

slots from IWLSridge() or IWLSCoxridge()

character family with value "gaussian" (also for "linear"), "binomial" (also for "logistic"), or "cox"

\(q\)-dimensional vector penalties containing optimised regularisation hyperparameters (one for each predictor group)

list indices with nfolds slots (one for each cross-validation fold), each containing the indices of the observations

list datablocks with \(q\) slots (one for each predictor group), each containing an \(n_0 \times p_k\) matrix, where \(k \in \{1, \ldots, q\}\)

\(p\)-dimensional group vector z (see argument)

list pars with slots family (see above), the \(p\)-dimensional vectors mu.x and sd.x and the scalars mu.y and sd.y

Details

The numbers of observations (samples) for training or testing are indicated by \(n_0\) and \(n_1\), respectively, the number of predictors (features) is indicated by \(p\), and the number of predictor group is indicated by \(q\). Observations are indexed by \(i\) in \(\{1, \ldots, n\}\), predictors are indexed by \(j\) in \(\{1, \ldots, p\}\), and predictor groups are indexed by \(k\) in \(\{1, \ldots, q\}\). The number of predictors in the \(k^{\text{th}}\) group is indicated by \(p_k\), with \(\sum_{k=1}^q p_k = p\) for non-overlapping groups.

References

Mark A. van de Wiel, Mirrelijn M. van Nee and Armin Rauschenberger (2021). "Fast cross-validation for multi-penalty high-dimensional ridge regression" Journal of Computational and Graphical Statistics 30(4):835-847. doi:10.1080/10618600.2021.1904962 .

See also

Examples

# minimal example
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
multiridge(x = x, y = y, z = z)
#> $etas
#>                [,1]
#>  [1,] -1.695214e-05
#>  [2,]  2.298572e-06
#>  [3,]  4.006165e-05
#>  [4,] -5.238749e-05
#>  [5,]  7.236052e-06
#>  [6,] -1.085261e-06
#>  [7,] -4.410569e-05
#>  [8,]  1.258423e-05
#>  [9,]  1.455452e-05
#> [10,]  2.074847e-05
#> [11,] -1.733551e-05
#> [12,] -3.033642e-05
#> [13,] -5.545538e-06
#> [14,]  2.428599e-05
#> [15,] -2.695604e-06
#> [16,]  7.123617e-06
#> [17,]  4.762505e-05
#> [18,]  4.544169e-06
#> [19,] -5.363664e-06
#> [20,] -5.635245e-06
#> [21,]  3.222958e-06
#> [22,] -2.345058e-05
#> [23,]  1.007538e-05
#> [24,]  2.415563e-05
#> [25,] -2.634678e-05
#> [26,]  1.733121e-05
#> [27,] -1.495599e-05
#> [28,]  3.633766e-05
#> [29,]  2.878147e-05
#> [30,]  3.225878e-06
#> [31,]  3.369244e-06
#> [32,] -2.619020e-05
#> [33,] -1.699597e-05
#> [34,]  2.283179e-05
#> [35,] -1.065297e-05
#> [36,]  1.871133e-05
#> [37,]  1.129965e-05
#> [38,]  2.110009e-05
#> [39,] -9.675909e-07
#> [40,]  1.914838e-05
#> [41,] -1.637166e-05
#> [42,] -1.462915e-05
#> [43,] -1.464945e-05
#> [44,]  1.415546e-05
#> [45,] -4.924877e-06
#> [46,] -4.178583e-05
#> [47,]  2.423454e-05
#> [48,] -4.588197e-05
#> [49,]  4.307152e-06
#> [50,] -4.104581e-06
#> 
#> $Ypred
#>                [,1]
#>  [1,] -1.695214e-05
#>  [2,]  2.298572e-06
#>  [3,]  4.006165e-05
#>  [4,] -5.238749e-05
#>  [5,]  7.236052e-06
#>  [6,] -1.085261e-06
#>  [7,] -4.410569e-05
#>  [8,]  1.258423e-05
#>  [9,]  1.455452e-05
#> [10,]  2.074847e-05
#> [11,] -1.733551e-05
#> [12,] -3.033642e-05
#> [13,] -5.545538e-06
#> [14,]  2.428599e-05
#> [15,] -2.695604e-06
#> [16,]  7.123617e-06
#> [17,]  4.762505e-05
#> [18,]  4.544169e-06
#> [19,] -5.363664e-06
#> [20,] -5.635245e-06
#> [21,]  3.222958e-06
#> [22,] -2.345058e-05
#> [23,]  1.007538e-05
#> [24,]  2.415563e-05
#> [25,] -2.634678e-05
#> [26,]  1.733121e-05
#> [27,] -1.495599e-05
#> [28,]  3.633766e-05
#> [29,]  2.878147e-05
#> [30,]  3.225878e-06
#> [31,]  3.369244e-06
#> [32,] -2.619020e-05
#> [33,] -1.699597e-05
#> [34,]  2.283179e-05
#> [35,] -1.065297e-05
#> [36,]  1.871133e-05
#> [37,]  1.129965e-05
#> [38,]  2.110009e-05
#> [39,] -9.675909e-07
#> [40,]  1.914838e-05
#> [41,] -1.637166e-05
#> [42,] -1.462915e-05
#> [43,] -1.464945e-05
#> [44,]  1.415546e-05
#> [45,] -4.924877e-06
#> [46,] -4.178583e-05
#> [47,]  2.423454e-05
#> [48,] -4.588197e-05
#> [49,]  4.307152e-06
#> [50,] -4.104581e-06
#> 
#> $convergence
#> [1] FALSE
#> 
#> $nIt
#> [1] 1
#> 
#> $Hresl
#> $Hresl$Hmatl
#>                [,1]
#>  [1,] -1.695214e-05
#>  [2,]  2.298572e-06
#>  [3,]  4.006165e-05
#>  [4,] -5.238749e-05
#>  [5,]  7.236052e-06
#>  [6,] -1.085261e-06
#>  [7,] -4.410569e-05
#>  [8,]  1.258423e-05
#>  [9,]  1.455452e-05
#> [10,]  2.074847e-05
#> [11,] -1.733551e-05
#> [12,] -3.033642e-05
#> [13,] -5.545538e-06
#> [14,]  2.428599e-05
#> [15,] -2.695604e-06
#> [16,]  7.123617e-06
#> [17,]  4.762505e-05
#> [18,]  4.544169e-06
#> [19,] -5.363664e-06
#> [20,] -5.635245e-06
#> [21,]  3.222958e-06
#> [22,] -2.345058e-05
#> [23,]  1.007538e-05
#> [24,]  2.415563e-05
#> [25,] -2.634678e-05
#> [26,]  1.733121e-05
#> [27,] -1.495599e-05
#> [28,]  3.633766e-05
#> [29,]  2.878147e-05
#> [30,]  3.225878e-06
#> [31,]  3.369244e-06
#> [32,] -2.619020e-05
#> [33,] -1.699597e-05
#> [34,]  2.283179e-05
#> [35,] -1.065297e-05
#> [36,]  1.871133e-05
#> [37,]  1.129965e-05
#> [38,]  2.110009e-05
#> [39,] -9.675909e-07
#> [40,]  1.914838e-05
#> [41,] -1.637166e-05
#> [42,] -1.462915e-05
#> [43,] -1.464945e-05
#> [44,]  1.415546e-05
#> [45,] -4.924877e-06
#> [46,] -4.178583e-05
#> [47,]  2.423454e-05
#> [48,] -4.588197e-05
#> [49,]  4.307152e-06
#> [50,] -4.104581e-06
#> 
#> $Hresl$Mmatl
#>              [,1]
#>  [1,] -0.86821837
#>  [2,]  1.48525400
#>  [3,]  1.70041507
#>  [4,] -1.78312014
#>  [5,]  2.02606606
#>  [6,] -1.45314210
#>  [7,] -0.64053155
#>  [8,]  0.52446488
#>  [9,]  0.75154961
#> [10,] -0.41295528
#> [11,] -0.02960591
#> [12,] -0.49399931
#> [13,]  0.38858070
#> [14,] -0.54482525
#> [15,] -0.71198117
#> [16,]  1.02743526
#> [17,]  0.67931190
#> [18,] -0.41383038
#> [19,] -0.62576224
#> [20,] -0.87286255
#> [21,] -0.58696016
#> [22,] -0.45482129
#> [23,]  0.59701666
#> [24,]  1.41738748
#> [25,]  0.10967369
#> [26,] -1.57248241
#> [27,] -0.75168391
#> [28,] -1.01735302
#> [29,]  0.05112213
#> [30,]  0.11591661
#> [31,]  0.87824567
#> [32,]  0.89982280
#> [33,] -0.89060253
#> [34,]  0.45010187
#> [35,] -1.23896227
#> [36,]  0.78421582
#> [37,] -0.18919110
#> [38,]  0.02417846
#> [39,] -0.06105716
#> [40,]  0.44444921
#> [41,]  0.70839106
#> [42,] -0.77983620
#> [43,] -1.95328639
#> [44,]  1.39229614
#> [45,] -0.67413493
#> [46,] -1.29205794
#> [47,]  2.50075838
#> [48,]  0.68186819
#> [49,] -0.09482760
#> [50,]  0.76956951
#> 
#> 
#> $linearized
#>              [,1]
#>  [1,] -0.86823532
#>  [2,]  1.48525630
#>  [3,]  1.70045513
#>  [4,] -1.78317252
#>  [5,]  2.02607329
#>  [6,] -1.45314318
#>  [7,] -0.64057565
#>  [8,]  0.52447746
#>  [9,]  0.75156416
#> [10,] -0.41293453
#> [11,] -0.02962324
#> [12,] -0.49402965
#> [13,]  0.38857516
#> [14,] -0.54480096
#> [15,] -0.71198387
#> [16,]  1.02744239
#> [17,]  0.67935952
#> [18,] -0.41382584
#> [19,] -0.62576760
#> [20,] -0.87286818
#> [21,] -0.58695693
#> [22,] -0.45484474
#> [23,]  0.59702674
#> [24,]  1.41741164
#> [25,]  0.10964734
#> [26,] -1.57246508
#> [27,] -0.75169887
#> [28,] -1.01731668
#> [29,]  0.05115091
#> [30,]  0.11591984
#> [31,]  0.87824903
#> [32,]  0.89979661
#> [33,] -0.89061953
#> [34,]  0.45012471
#> [35,] -1.23897293
#> [36,]  0.78423453
#> [37,] -0.18917980
#> [38,]  0.02419956
#> [39,] -0.06105813
#> [40,]  0.44446836
#> [41,]  0.70837469
#> [42,] -0.77985083
#> [43,] -1.95330104
#> [44,]  1.39231029
#> [45,] -0.67413986
#> [46,] -1.29209973
#> [47,]  2.50078262
#> [48,]  0.68182231
#> [49,] -0.09482329
#> [50,]  0.76956540
#> 
#> $unpen
#> [1] FALSE
#> 
#> $intercept
#> [1] TRUE
#> 
#> $eta0
#> [1] 0
#> 
#> $X1
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1
#> [31,]    1
#> [32,]    1
#> [33,]    1
#> [34,]    1
#> [35,]    1
#> [36,]    1
#> [37,]    1
#> [38,]    1
#> [39,]    1
#> [40,]    1
#> [41,]    1
#> [42,]    1
#> [43,]    1
#> [44,]    1
#> [45,]    1
#> [46,]    1
#> [47,]    1
#> [48,]    1
#> [49,]    1
#> [50,]    1
#> 
#> $family
#> [1] "gaussian"
#> 
#> $penalties
#> [1] 1795940.5 2032905.4  990739.8  635402.1 1832878.1
#> 
#> $datablocks
#> $datablocks[[1]]
#>               [,1]        [,2]         [,3]        [,4]
#>  [1,]  0.011955271  0.45897301 -0.476277928  1.17302205
#>  [2,]  1.341961542 -0.77861421 -0.187519827 -2.34507685
#>  [3,]  0.084946866  0.60587986 -0.983272521 -0.76258294
#>  [4,]  1.214296541 -2.05552744 -1.993502812  0.52463550
#>  [5,]  0.545698387  0.04787017  0.174407520 -1.26993670
#>  [6,]  1.564848768 -1.10214113  0.355995716  1.05880703
#>  [7,]  0.204052609  0.78545059  2.037168843  1.40317815
#>  [8,] -0.537916421  0.04596475 -1.789716957 -0.60348129
#>  [9,] -2.871159476  1.05576421  0.094254008 -0.33325108
#> [10,] -1.132649855 -0.04821903 -0.771800703  1.08001639
#> [11,] -1.563586234 -0.68247660  0.387018160 -0.22450293
#> [12,] -0.986661874  0.46876358 -0.345840958 -0.50255102
#> [13,]  0.167010279 -0.16110627 -1.230971219  2.27597715
#> [14,] -0.832536711  1.23653611 -1.130900436  0.35503971
#> [15,]  0.477911077  0.21538578 -0.745634795  0.18665328
#> [16,]  1.034358233  0.45117992  0.640839380  0.08775603
#> [17,] -0.589192541  0.37089718 -0.259111048 -0.46550719
#> [18,] -1.830879301 -1.06310558 -0.551971041  0.64527493
#> [19,] -1.094135006 -0.80384476  1.335212206 -1.56430909
#> [20,]  1.807551399 -1.75561628  1.905362742 -0.33682459
#> [21,] -0.477515487  0.30220844 -1.016075470 -1.38580529
#> [22,]  0.117307455  0.76390511 -0.844906722  0.18202848
#> [23,] -2.463512872  2.18420163  0.751696654  0.62330922
#> [24,]  0.002801761 -0.80120419 -1.899314741 -0.77785207
#> [25,] -0.054568879 -0.64534918 -0.349086607  0.29537068
#> [26,]  0.618630310  1.56535267  0.777938590 -2.39361834
#> [27,]  0.027122935  1.16434859  0.688628096  0.18672149
#> [28,] -0.586371522 -0.50818067  0.460923334 -0.20506936
#> [29,] -0.017949132  0.86616215 -0.367498202 -0.90165685
#> [30,]  1.202894733 -0.66483719  0.028073739 -1.35858220
#> [31,]  0.438883350  1.54263864  0.602545065  0.55662483
#> [32,]  0.062495284  0.33093597 -0.344934116  1.04684951
#> [33,]  0.337982858 -0.33744570  0.102712152 -0.17938566
#> [34,] -1.111422434 -1.58898303  0.777269335  1.55348350
#> [35,]  0.035568523 -1.51555689  0.860133069  0.26497492
#> [36,] -1.260918372 -0.96668904 -0.004622434 -1.49954334
#> [37,] -0.172315209  1.15702401  1.822658650 -0.30945003
#> [38,]  0.334816586 -1.61538986  0.501210532  0.81855097
#> [39,]  0.983498555 -0.70122702 -1.957034851  0.68651699
#> [40,]  1.129544409  1.46976510  0.365071997  0.78710823
#> [41,] -0.078481419  0.50012186 -0.404971989  0.85434392
#> [42,]  0.826571947  0.61755006  1.627044219  0.24068430
#> [43,] -0.079561689  0.31455792 -0.851958176  0.05361412
#> [44,]  0.546855533  0.08566411  1.001906091 -1.19486602
#> [45,] -0.658888124  1.43663722 -1.029770284  1.41273759
#> [46,]  1.834125124  0.94510593  0.204145228 -0.75813740
#> [47,]  0.283949326 -0.72050628 -0.064198689  0.13824649
#> [48,]  0.050902417 -0.48551835 -0.405927373  1.17194661
#> [49,]  0.494019015 -0.93572018  1.055928705 -0.55072893
#> [50,]  0.617661467 -1.05158569  1.448675868  0.25924709
#> 
#> $datablocks[[2]]
#>              [,1]         [,2]        [,3]        [,4]
#>  [1,] -0.64486070 -1.630323078  1.50755925  0.03577889
#>  [2,]  0.32024756 -0.251853823 -0.42032445  0.91487547
#>  [3,]  0.26531225 -0.403668035  0.49916746  1.08516818
#>  [4,] -0.71627699  1.684755636 -0.87008402 -0.80780673
#>  [5,]  1.16082649  1.459350004 -0.31295289  0.53403865
#>  [6,]  0.13588938  0.039543646  0.56547688 -0.91747149
#>  [7,]  0.14716193  0.902213807 -0.53563637 -0.28677230
#>  [8,]  0.06892991  0.613200070 -1.12147662 -0.54584986
#>  [9,] -0.05471740 -0.660181103 -0.10775332  2.03651008
#> [10,] -0.49386401  1.663014292 -1.89275483 -0.19447901
#> [11,] -0.94019853  0.600963975 -0.19055627 -1.15903100
#> [12,]  0.48493961  0.048204258  0.74546076 -0.76866559
#> [13,]  1.71445247  0.085510476  0.92266441  0.44120376
#> [14,]  1.49176111  0.303094362  1.44030280 -0.55844269
#> [15,]  0.48599608 -0.870900710 -1.12394254  0.01887352
#> [16,] -1.30896262  0.111286187  1.05104893  0.34919573
#> [17,]  1.48760012  0.809487937 -0.69438150 -0.83142009
#> [18,]  1.45259625 -0.461776709 -2.43518965  0.50891526
#> [19,]  0.20412768  0.013444787  0.11898629  0.75467135
#> [20,] -2.86244720  0.036942629 -0.48720779 -2.35520835
#> [21,] -0.86836174  0.488381823  0.77130206 -0.83196136
#> [22,]  0.48505076  0.181710424 -0.74696599 -0.27862761
#> [23,] -0.42694769 -1.855918818 -1.21542541 -0.95812175
#> [24,] -1.08631703 -1.619805198  0.53656962  0.04249157
#> [25,]  0.25747609  1.515959878  2.27866613  0.35844270
#> [26,]  0.73231510 -0.348133126 -0.04952387 -0.20588343
#> [27,]  0.86235563  0.464996372 -1.24622810 -0.53235768
#> [28,]  0.70413290 -0.811775054  1.75399907  1.79763588
#> [29,] -0.38768765  1.502909193  0.75741521  1.84527905
#> [30,] -0.52626996 -1.336299623 -1.14905455 -1.24409885
#> [31,] -0.20423700 -2.175730629  0.58794213 -0.60043822
#> [32,]  0.86257253 -0.560986555 -1.20148679 -1.65769574
#> [33,] -1.44964537  0.504121135 -0.33643926  0.10102143
#> [34,]  0.72416393  0.174725573  0.76589208  1.50253761
#> [35,]  0.29358798 -1.547800094  0.35454665  0.95739520
#> [36,] -2.20391686 -0.455736272  1.20922286  0.09495945
#> [37,] -0.02453426 -0.007632111  0.20816125  1.32384403
#> [38,] -0.40271969  1.259656370  0.21822166 -0.19592183
#> [39,]  0.90563147  2.242905460  0.73613345 -1.16088981
#> [40,] -0.88151324 -0.921238505 -1.52867543  0.04131048
#> [41,]  1.37390187  0.619154032  0.85405052 -0.66048593
#> [42,] -1.01459905  0.479162949  0.02916112  0.38608718
#> [43,] -0.90661061  0.443148603 -0.09193009 -0.15942035
#> [44,] -1.29318837 -0.089358459 -0.44875321 -2.07483431
#> [45,] -0.20884664 -1.019429851 -0.85995853  1.02250950
#> [46,] -0.02226225  1.001900001  0.97892919  0.81000141
#> [47,]  1.33450002 -0.120447158 -1.08406807  2.06093673
#> [48,]  1.34944484 -1.529263473  0.46140357  0.22515873
#> [49,] -1.08329239 -0.596233531 -0.44054178 -0.51159530
#> [50,]  0.70730329  0.024748035  1.23902800  0.24863743
#> 
#> $datablocks[[3]]
#>               [,1]        [,2]       [,3]        [,4]
#>  [1,] -0.664729181 -0.75240978  0.8131966  0.96500229
#>  [2,] -0.049679268  0.26978974 -0.5751792 -0.06640777
#>  [3,] -0.577836721  0.68751101 -0.2924420 -0.29259583
#>  [4,]  0.764659984 -1.87999969 -1.3666506  0.54550878
#>  [5,]  0.312197520 -0.94897569 -1.1323618 -0.03247371
#>  [6,]  1.615056535  1.03226905 -1.3605927 -1.33013315
#>  [7,] -1.380991857 -0.71349462  0.1006333  0.29716962
#>  [8,] -1.800603426  0.53469523  1.0837938 -0.54449706
#>  [9,] -1.147433231  1.92129392  0.6874826 -0.44936194
#> [10,]  0.759916018  1.90516662  0.4101414  1.65904817
#> [11,] -0.276087658 -0.23105501 -1.5330130  0.63222859
#> [12,]  0.475379160  0.11926135  0.3018096 -0.56723732
#> [13,]  0.814006088 -0.64376242  0.5398868 -0.19341783
#> [14,] -0.110375198  0.42840374  0.5666505  0.90410457
#> [15,]  0.452114934 -0.69512705 -1.1987824  1.20419186
#> [16,] -0.231896902  1.08178547 -1.6033388 -0.57243393
#> [17,] -1.045212088  0.74171893  1.0888413  1.76080541
#> [18,] -0.981169224 -1.13863753  1.8906015 -2.84852493
#> [19,] -1.239635268 -1.65013078 -0.1090888  0.56664425
#> [20,]  0.958753183  1.54833398  0.7976199 -1.20334782
#> [21,] -0.063144091  1.15073952  1.0616077  1.87788219
#> [22,]  1.935598050 -1.29048264 -1.3924990 -0.17294096
#> [23,]  0.962420245  0.35140935 -0.5621199 -0.37985137
#> [24,]  0.187499103 -0.61788478  0.9106718 -0.42205671
#> [25,]  0.751922120 -1.38791801 -0.7192887  1.31502585
#> [26,] -1.585259463 -0.14719808  0.9569294  0.03206060
#> [27,] -2.342533490 -0.20346507 -0.4197090  0.78525718
#> [28,] -0.009475258  0.46640280  0.8280113 -0.48943495
#> [29,]  0.806147570 -0.66112718  0.0900955 -0.30733909
#> [30,]  0.950409558  0.67870930  1.1461989  1.01723660
#> [31,]  0.047442793 -0.59128621 -0.5582178 -0.18128526
#> [32,]  0.635996925 -0.07726423 -0.4105761 -1.95474423
#> [33,]  0.286634722 -0.94117577  0.8228294 -1.85316434
#> [34,] -0.574147573  1.01796131  0.6291789 -1.59110527
#> [35,] -0.983689229 -1.45845980  0.6302679  0.18240844
#> [36,] -2.874139866 -0.05978207  0.9021110 -0.90947071
#> [37,]  0.685902827  0.54707308  0.7502317  0.70347778
#> [38,]  0.293996128  0.26404248  1.4884916  0.14079028
#> [39,]  0.389094836  1.62099584 -0.8539365 -0.03848675
#> [40,] -1.202417227  0.47591762  0.2723361 -0.67450577
#> [41,]  1.432218110  0.33891866 -0.5197317 -1.01411046
#> [42,]  0.044081702 -1.89162312 -0.4284664 -0.49550143
#> [43,] -0.095131097 -0.22755610 -1.1166443  0.69869089
#> [44,]  0.805875540 -0.16770699  1.0652832  1.56807319
#> [45,]  0.696139497  1.86736177 -1.2488390 -0.12753606
#> [46,]  0.164875683  0.23954785 -2.8708284  0.24566072
#> [47,]  0.189518280 -0.80954024  1.1200599  0.95877754
#> [48,]  0.566171019 -0.44299160  0.2341789  0.43356342
#> [49,]  0.014889220  1.29421126 -0.6652459 -0.77387909
#> [50,]  1.236669965 -0.95446543 -0.2515886  0.99223552
#> 
#> $datablocks[[4]]
#>              [,1]         [,2]        [,3]         [,4]
#>  [1,]  0.61425873 -0.060600140 -0.22908160 -1.554875514
#>  [2,] -0.81038115  0.641640148  2.74697501  1.160562814
#>  [3,] -0.87832893  0.569447928 -1.12450564  0.605070147
#>  [4,]  0.69147573 -1.099169026  0.77200222  0.869559161
#>  [5,]  0.55565586  0.111047873 -0.50059084  0.877247587
#>  [6,]  0.27824187  1.439927570  0.55079990  1.431677931
#>  [7,]  0.36667355 -2.240352363 -0.44211064 -0.452458194
#>  [8,] -0.38476545 -0.408587429 -0.32996329 -0.835780617
#>  [9,] -0.67753686 -2.643680841 -0.23278792 -0.208269727
#> [10,]  1.43116671  2.531512923  0.27053363  0.780802366
#> [11,]  0.53250605  0.582303824 -0.16711135  0.274971998
#> [12,]  0.36008271 -0.521590813  1.79507031 -0.615578106
#> [13,]  0.44447052  0.057714981 -0.02573137 -0.646546358
#> [14,] -0.34330344  0.845138951 -1.43628558 -1.959148088
#> [15,]  0.06840580 -0.043293956 -0.16985463  0.353530317
#> [16,] -2.21405941 -0.050805484 -0.95841855 -1.088335049
#> [17,] -1.66016670  1.998330501 -0.98348016 -1.072955026
#> [18,]  1.22374873 -0.526993655  0.79969368  0.433904507
#> [19,] -1.74618162 -1.036537628  0.57244396 -0.157244080
#> [20,]  0.43384126  0.841646555  0.12935430  0.319020907
#> [21,]  0.44262522  0.125530667 -0.32656806 -2.052566905
#> [22,]  0.39106914 -0.194890722 -0.09413393 -0.707153138
#> [23,] -1.44617037  0.381101272  0.81919298 -0.516914658
#> [24,]  0.29744639  1.097608581 -0.46082241  0.004194602
#> [25,]  1.01470827  0.637390987 -0.60716033 -2.358237203
#> [26,]  0.28721529  0.322992010 -0.15018860  0.053590999
#> [27,] -1.47439025 -0.009056576  0.96406774 -0.939574753
#> [28,] -1.80135998 -1.373330236  0.46543491  1.796584682
#> [29,] -0.76218264 -0.043212542 -1.39638416  0.699639593
#> [30,]  1.48994002  0.436416617  0.95005992  1.720998969
#> [31,]  0.04886340  0.739121044 -0.63998152  0.052262660
#> [32,]  0.62644776  0.530654408  1.80537406  0.256234546
#> [33,]  0.73307349  0.587473519  0.72023585 -0.409885465
#> [34,] -1.16835222 -0.429765406 -0.68280262  0.287452985
#> [35,]  1.59405740  0.576467303 -0.56067335 -1.390562495
#> [36,] -0.44502915  0.278120542 -0.09665304  0.911666416
#> [37,] -0.49726867 -0.581466751  0.09976913 -0.743671460
#> [38,] -0.23938818  0.924374224 -0.69696254  0.243064671
#> [39,]  1.64607174 -1.273255110 -2.41637461  1.685194746
#> [40,]  0.48074783 -0.045992463 -2.38365536 -0.243715020
#> [41,]  0.20753050 -1.778522926  0.19717487  1.381863152
#> [42,] -0.20000254 -0.339171384 -0.29415411 -0.677539974
#> [43,] -1.60173083 -0.216907243  0.21529536 -0.403325195
#> [44,] -0.42005745  1.834482949  0.53450579 -0.034701602
#> [45,] -0.85027181 -0.883931633  1.33537472  0.605795113
#> [46,]  1.31818629 -1.142286600  0.76552722  1.689257843
#> [47,] -0.06300112 -0.969666665 -0.91847992  0.862002676
#> [48,]  2.15311364 -0.482410474  2.05591154 -0.914533733
#> [49,] -0.43935375 -0.307651492 -0.25134959  0.415004660
#> [50,]  0.39165864  0.612684182  0.01146858  0.212416313
#> 
#> $datablocks[[5]]
#>              [,1]        [,2]         [,3]        [,4]
#>  [1,]  0.59609103  0.90402687  1.340147799  0.62631156
#>  [2,] -0.19300135 -0.64413536 -0.593911174  0.13200343
#>  [3,] -1.21661681  0.29761661  0.529115236 -0.18854143
#>  [4,]  1.00552453 -1.34890756  0.370433364  0.34328984
#>  [5,]  0.24569217 -0.18533964 -1.055675107  2.44505150
#>  [6,] -0.72458305 -0.85933005  0.416460352 -0.50869399
#>  [7,]  0.92266530  2.16208506  0.688466460 -0.77080121
#>  [8,]  1.06205651  0.38527482 -1.081801353  0.69092195
#>  [9,]  1.94557725 -1.40535945 -1.196948004  0.06323410
#> [10,]  0.43988151  2.63574649 -0.101038504  2.01605810
#> [11,] -0.09008307 -1.02426284  0.276067035 -0.94129142
#> [12,] -1.29747004 -0.93182964  2.585368818 -1.05226804
#> [13,] -0.54192982  0.23814460  0.080202497 -0.51423463
#> [14,] -1.35586889  0.12018539 -0.617732541 -0.47500058
#> [15,]  0.90187509 -0.69141615 -1.256665051  0.02488599
#> [16,] -0.47419524 -1.26347898 -0.225352581  0.32696382
#> [17,] -0.72684802 -0.89089920  0.805883743 -1.12890203
#> [18,]  0.47827571  0.14479601  0.591375406  0.54173134
#> [19,]  1.45362416  0.08467908  0.728510436 -0.55025490
#> [20,] -0.57493599  0.63403185  0.134299558  0.87506544
#> [21,] -0.51364717 -0.82988894 -0.807275274  2.13162994
#> [22,]  0.65209189  0.64135147 -0.310935599 -0.05581880
#> [23,] -0.37473012  0.47368913 -0.363558817 -0.28147983
#> [24,]  1.85927842 -1.28413755 -0.032562516  0.47791109
#> [25,] -1.16610136 -1.55527479 -0.501958739  0.93459840
#> [26,] -0.07258665  0.37711765  0.164075564 -0.20845915
#> [27,]  1.21743964 -1.38139414  0.900911008 -0.75710989
#> [28,] -0.09207967  1.57832193 -1.114192185  0.75003664
#> [29,] -0.85087208  1.47571641  0.470503824  0.32346002
#> [30,] -0.50901658 -1.26517597  0.689796333 -0.56413561
#> [31,]  0.44201856 -0.02002963  0.556424454 -1.46596347
#> [32,]  0.74974524  1.22869362 -0.699338253 -2.11923092
#> [33,]  0.79842201  1.05786498  0.963885940  0.35630970
#> [34,] -0.27917462 -0.23360360  2.022094142 -0.87572046
#> [35,] -0.09244283 -0.29091225 -0.015865570 -1.04196257
#> [36,]  0.46387667  0.41035734  0.731199683  0.08152964
#> [37,] -1.88112434 -1.79000956 -0.032605219  1.66435294
#> [38,] -1.25592182  1.37592914 -0.848182319 -1.07713121
#> [39,]  2.71655123 -0.75413075 -1.972683832 -0.55678897
#> [40,] -0.61128481  0.43526642  1.054472013  1.55024172
#> [41,]  0.62181844  0.63350978 -0.003747501 -0.89378955
#> [42,] -2.20987252  0.47442818 -3.208716326 -0.20163116
#> [43,]  0.08877606 -0.45503261  1.154458959 -2.03329593
#> [44,] -1.18509002  0.42146813  0.711551254  1.14195615
#> [45,] -0.13463833 -0.54027466  0.651716907  0.86366404
#> [46,] -0.11808820  0.58223221 -1.487315105 -0.20336215
#> [47,] -0.48876985  0.75268603 -0.609000848 -0.64977326
#> [48,]  0.41779615  0.97177368  0.431910646  0.60779117
#> [49,]  0.89518159 -0.29514177 -0.481012875  0.89945241
#> [50,] -0.94328591 -0.55702782 -0.431256136 -0.75280976
#> 
#> 
#> $indices
#> $indices[[1]]
#> [1]  4 13 21 46 50
#> 
#> $indices[[2]]
#> [1]  5  7 18 20 26
#> 
#> $indices[[3]]
#> [1] 10 28 32 33 44
#> 
#> $indices[[4]]
#> [1] 12 17 27 31 34
#> 
#> $indices[[5]]
#> [1]  8 16 29 38 42
#> 
#> $indices[[6]]
#> [1]  1  3  9 23 24
#> 
#> $indices[[7]]
#> [1] 25 30 37 40 41
#> 
#> $indices[[8]]
#> [1]  6 11 14 45 48
#> 
#> $indices[[9]]
#> [1]  2 15 22 35 49
#> 
#> $indices[[10]]
#> [1] 19 36 39 43 47
#> 
#> 
#> $z
#>  [1] 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5
#> 
#> $pars
#> $pars$family
#> [1] "gaussian"
#> 
#> $pars$mu.x
#>  [1] -0.2193549478 -0.1236907683 -0.0281794890  0.1049095958  0.1439363980
#>  [6]  0.0260020961  0.3176702862  0.0041437583 -0.0504108085  0.1149841094
#> [11]  0.1098525680 -0.2999377666  0.0278264294 -0.0026918626 -0.0928262058
#> [16] -0.0732010983 -0.0003578349 -0.1098120889  0.0728249145  0.0334349542
#> 
#> $pars$sd.x
#>  [1] 0.9185759 1.0428025 0.9456351 1.0413831 1.0768972 0.9600772 0.9766040
#>  [8] 0.9731598 0.8931155 1.2249541 0.8394730 1.0641089 0.8939175 0.7883511
#> [15] 1.0027223 1.1073157 0.8374548 1.0623911 0.8822887 0.9850965
#> 
#> $pars$mu.y
#> [1] 0.01537418
#> 
#> $pars$sd.y
#> [1] 1.075638
#> 
#> 
#> attr(,"class")
#> [1] "multiridge"

# \donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x %*% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
#>     glmnet multiridge 
#>  0.4692074  0.5982584 
sapply(coef, function(x) mean((beta-x[-1])^2))
#>     glmnet multiridge 
#>  0.1718059  0.1262343 

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
#>     glmnet multiridge 
#>   73.65774   50.10826 
# }