Multi-penalty ridge regression

Fits multi-penalty ridge regression (tuning regularisation hyperparameters and estimating regression coefficients). This is a wrapper function of some functions from the multiridge-package.

Usage

multiridge(
  x,
  y,
  z,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10,
  penalties = NULL
)

Arguments

x

predictors: \(n \times p\) matrix

y

response: \(n\)-dimensional vector

z

\(p\)-dimensional vector with entries in \(\{1, \ldots, q\}\)

family

character "linear" (or "gaussian"), "logistic" (or "binomial"), or "cox"

foldid

\(n_0\)-dimensional vector containing the fold identifiers

nfolds

integer specifying the number of folds

penalties

\(q\)-dimensional vector of penalty parameters, or NULL (cross-validation)

Value

Returns an object of class "multiridge", a list with the following slots:

slots from IWLSridge() or IWLSCoxridge()

character family with value "gaussian" (also for "linear"), "binomial" (also for "logistic"), or "cox"

\(q\)-dimensional vector penalties containing optimised regularisation hyperparameters (one for each predictor group)

list indices with nfolds slots (one for each cross-validation fold), each containing the indices of the observations

list datablocks with \(q\) slots (one for each predictor group), each containing an \(n_0 \times p_k\) matrix, where \(k \in \{1, \ldots, q\}\)

\(p\)-dimensional group vector z (see argument)

list pars with slots family (see above), the \(p\)-dimensional vectors mu.x and sd.x and the scalars mu.y and sd.y

Details

The number of observations (samples) for training or testing are indicated by \(n_0\) and \(n_1\), respectively, the number of variables (features) is indicated by \(p\), and the number of variable groups is indicated by \(q\). Observations (samples) are indexed by \(i\) in \(\{1, \ldots, n\}\), variables (features) are indexed by \(j\) in \(\{1, \ldots, p\}\), and variable groups are indexed by \(k\) in \(\{1, \ldots, q\}\). The number of variables in the \(k^{\text{th}}\) group is indicated by \(p_k\), with \(\sum_{k=1}^q p_k = p\) for non-overlapping groups.

References

Mark A. van de Wiel, Mirrelijn M. van Nee and Armin Rauschenberger (2021). "Fast cross-validation for multi-penalty high-dimensional ridge regression" Journal of Computational and Graphical Statistics 30(4):835-847. doi: 10.1080/10618600.2021.1904962. # nolint: line_length_linter.

See also

Examples

# minimal example
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
multiridge(x = x, y = y, z = z)
#> $etas
#>                [,1]
#>  [1,]  0.0789277961
#>  [2,] -0.1158981644
#>  [3,]  0.1989534540
#>  [4,]  0.0542538405
#>  [5,] -0.0025777485
#>  [6,] -0.2032923669
#>  [7,]  0.0051482353
#>  [8,]  0.0936596650
#>  [9,] -0.0519364930
#> [10,] -0.0594134344
#> [11,] -0.1461481221
#> [12,]  0.0006567058
#> [13,]  0.0478645692
#> [14,]  0.0390248724
#> [15,] -0.0131573252
#> [16,] -0.0814534869
#> [17,] -0.1238979853
#> [18,]  0.0756429386
#> [19,] -0.0059500787
#> [20,]  0.3140044256
#> [21,] -0.2109483105
#> [22,] -0.2266181161
#> [23,] -0.0179444214
#> [24,]  0.0963217058
#> [25,] -0.0439431345
#> [26,]  0.0373154516
#> [27,]  0.2610273662
#> [28,]  0.0505866003
#> [29,] -0.0214941244
#> [30,]  0.3650935765
#> [31,] -0.0036548692
#> [32,] -0.0483599274
#> [33,] -0.1097776003
#> [34,]  0.1111562156
#> [35,] -0.0326976689
#> [36,] -0.2670942657
#> [37,] -0.0463956862
#> [38,]  0.1210370975
#> [39,]  0.3019248720
#> [40,]  0.0946900111
#> [41,] -0.1565984827
#> [42,]  0.0961801806
#> [43,] -0.2053656848
#> [44,] -0.1355743931
#> [45,]  0.2146563467
#> [46,] -0.2922205187
#> [47,]  0.1183390123
#> [48,]  0.1371129507
#> [49,]  0.0062342179
#> [50,] -0.2973996981
#> 
#> $Ypred
#>                [,1]
#>  [1,]  0.0789277961
#>  [2,] -0.1158981644
#>  [3,]  0.1989534540
#>  [4,]  0.0542538405
#>  [5,] -0.0025777485
#>  [6,] -0.2032923669
#>  [7,]  0.0051482353
#>  [8,]  0.0936596650
#>  [9,] -0.0519364930
#> [10,] -0.0594134344
#> [11,] -0.1461481221
#> [12,]  0.0006567058
#> [13,]  0.0478645692
#> [14,]  0.0390248724
#> [15,] -0.0131573252
#> [16,] -0.0814534869
#> [17,] -0.1238979853
#> [18,]  0.0756429386
#> [19,] -0.0059500787
#> [20,]  0.3140044256
#> [21,] -0.2109483105
#> [22,] -0.2266181161
#> [23,] -0.0179444214
#> [24,]  0.0963217058
#> [25,] -0.0439431345
#> [26,]  0.0373154516
#> [27,]  0.2610273662
#> [28,]  0.0505866003
#> [29,] -0.0214941244
#> [30,]  0.3650935765
#> [31,] -0.0036548692
#> [32,] -0.0483599274
#> [33,] -0.1097776003
#> [34,]  0.1111562156
#> [35,] -0.0326976689
#> [36,] -0.2670942657
#> [37,] -0.0463956862
#> [38,]  0.1210370975
#> [39,]  0.3019248720
#> [40,]  0.0946900111
#> [41,] -0.1565984827
#> [42,]  0.0961801806
#> [43,] -0.2053656848
#> [44,] -0.1355743931
#> [45,]  0.2146563467
#> [46,] -0.2922205187
#> [47,]  0.1183390123
#> [48,]  0.1371129507
#> [49,]  0.0062342179
#> [50,] -0.2973996981
#> 
#> $convergence
#> [1] FALSE
#> 
#> $nIt
#> [1] 1
#> 
#> $Hresl
#> $Hresl$Hmatl
#>                [,1]
#>  [1,]  0.0789277961
#>  [2,] -0.1158981644
#>  [3,]  0.1989534540
#>  [4,]  0.0542538405
#>  [5,] -0.0025777485
#>  [6,] -0.2032923669
#>  [7,]  0.0051482353
#>  [8,]  0.0936596650
#>  [9,] -0.0519364930
#> [10,] -0.0594134344
#> [11,] -0.1461481221
#> [12,]  0.0006567058
#> [13,]  0.0478645692
#> [14,]  0.0390248724
#> [15,] -0.0131573252
#> [16,] -0.0814534869
#> [17,] -0.1238979853
#> [18,]  0.0756429386
#> [19,] -0.0059500787
#> [20,]  0.3140044256
#> [21,] -0.2109483105
#> [22,] -0.2266181161
#> [23,] -0.0179444214
#> [24,]  0.0963217058
#> [25,] -0.0439431345
#> [26,]  0.0373154516
#> [27,]  0.2610273662
#> [28,]  0.0505866003
#> [29,] -0.0214941244
#> [30,]  0.3650935765
#> [31,] -0.0036548692
#> [32,] -0.0483599274
#> [33,] -0.1097776003
#> [34,]  0.1111562156
#> [35,] -0.0326976689
#> [36,] -0.2670942657
#> [37,] -0.0463956862
#> [38,]  0.1210370975
#> [39,]  0.3019248720
#> [40,]  0.0946900111
#> [41,] -0.1565984827
#> [42,]  0.0961801806
#> [43,] -0.2053656848
#> [44,] -0.1355743931
#> [45,]  0.2146563467
#> [46,] -0.2922205187
#> [47,]  0.1183390123
#> [48,]  0.1371129507
#> [49,]  0.0062342179
#> [50,] -0.2973996981
#> 
#> $Hresl$Mmatl
#>              [,1]
#>  [1,]  0.55673511
#>  [2,]  1.04432598
#>  [3,]  0.49454337
#>  [4,] -0.73338057
#>  [5,]  0.61786339
#>  [6,] -0.60171945
#>  [7,]  0.84755309
#>  [8,]  1.01076711
#>  [9,]  1.47991093
#> [10,] -0.38358722
#> [11,] -0.68360468
#> [12,] -0.49600284
#> [13,] -0.25096949
#> [14,] -1.04748116
#> [15,] -2.01771504
#> [16,]  0.24464162
#> [17,]  0.47086471
#> [18,]  1.97277587
#> [19,] -1.81867829
#> [20,] -0.23193664
#> [21,] -0.58348530
#> [22,]  0.60498148
#> [23,] -0.34539136
#> [24,] -1.35546453
#> [25,] -1.11392181
#> [26,] -0.80526753
#> [27,]  0.37421891
#> [28,] -0.32614630
#> [29,] -0.55045816
#> [30,]  0.97290132
#> [31,]  1.91867768
#> [32,] -0.99329499
#> [33,] -0.75864387
#> [34,]  0.63628454
#> [35,] -1.90285049
#> [36,] -0.09952631
#> [37,]  0.82039492
#> [38,]  0.56257434
#> [39,]  0.15123514
#> [40,] -0.47994427
#> [41,]  0.17168770
#> [42,]  0.50030986
#> [43,] -0.15705231
#> [44,]  0.22620235
#> [45,]  0.55866556
#> [46,]  1.14940569
#> [47,] -0.13634036
#> [48,]  1.69420816
#> [49,]  0.48769896
#> [50,] -1.69656482
#> 
#> 
#> $linearized
#>              [,1]
#>  [1,]  0.63566290
#>  [2,]  0.92842782
#>  [3,]  0.69349682
#>  [4,] -0.67912673
#>  [5,]  0.61528565
#>  [6,] -0.80501182
#>  [7,]  0.85270132
#>  [8,]  1.10442677
#>  [9,]  1.42797443
#> [10,] -0.44300066
#> [11,] -0.82975280
#> [12,] -0.49534613
#> [13,] -0.20310492
#> [14,] -1.00845629
#> [15,] -2.03087237
#> [16,]  0.16318814
#> [17,]  0.34696672
#> [18,]  2.04841881
#> [19,] -1.82462837
#> [20,]  0.08206778
#> [21,] -0.79443361
#> [22,]  0.37836337
#> [23,] -0.36333578
#> [24,] -1.25914282
#> [25,] -1.15786495
#> [26,] -0.76795208
#> [27,]  0.63524628
#> [28,] -0.27555970
#> [29,] -0.57195228
#> [30,]  1.33799490
#> [31,]  1.91502281
#> [32,] -1.04165492
#> [33,] -0.86842147
#> [34,]  0.74744076
#> [35,] -1.93554816
#> [36,] -0.36662058
#> [37,]  0.77399923
#> [38,]  0.68361144
#> [39,]  0.45316002
#> [40,] -0.38525426
#> [41,]  0.01508922
#> [42,]  0.59649004
#> [43,] -0.36241800
#> [44,]  0.09062795
#> [45,]  0.77332190
#> [46,]  0.85718517
#> [47,] -0.01800134
#> [48,]  1.83132111
#> [49,]  0.49393317
#> [50,] -1.99396451
#> 
#> $unpen
#> [1] FALSE
#> 
#> $intercept
#> [1] TRUE
#> 
#> $eta0
#> [1] 0
#> 
#> $X1
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1
#> [31,]    1
#> [32,]    1
#> [33,]    1
#> [34,]    1
#> [35,]    1
#> [36,]    1
#> [37,]    1
#> [38,]    1
#> [39,]    1
#> [40,]    1
#> [41,]    1
#> [42,]    1
#> [43,]    1
#> [44,]    1
#> [45,]    1
#> [46,]    1
#> [47,]    1
#> [48,]    1
#> [49,]    1
#> [50,]    1
#> 
#> $family
#> [1] "gaussian"
#> 
#> $penalties
#> [1] 1.055143e+06 1.346826e+03 9.013287e+05 5.403942e+01 9.308759e+05
#> 
#> $datablocks
#> $datablocks[[1]]
#>                [,1]        [,2]         [,3]        [,4]
#>  [1,]  1.7704849193  0.58134615  1.501625117 -0.45646890
#>  [2,] -0.3136128776  1.25627928  1.545323960  0.90406442
#>  [3,]  0.6205141723  1.83001483 -0.071157211 -2.22025621
#>  [4,]  1.2375544129 -0.78175494 -0.079919493  0.31593692
#>  [5,]  1.3477724549  0.62498374 -1.102902099 -1.08957170
#>  [6,]  0.7272031432 -0.15969931  2.153006513  0.06943468
#>  [7,] -0.1128201976 -0.18419848 -0.030126246  0.08769511
#>  [8,] -1.8877528542  1.25619648 -1.061440213 -0.32129313
#>  [9,] -0.8771363857 -0.05208124  0.408483908  0.67895826
#> [10,] -2.7086155419  0.33730376  0.799918275  1.20567718
#> [11,]  0.8142954780  0.04846274 -0.477969405 -0.82277932
#> [12,] -0.2321841379 -0.52250063  0.037690775  0.54862628
#> [13,] -2.5134505971 -0.27468724  1.275119635 -0.66872923
#> [14,] -0.6503518563  0.33638909  0.105601662  0.69313155
#> [15,] -0.0937494156  1.63850882  1.156340980 -1.92476788
#> [16,] -0.0672181105 -0.78641825  0.534281769  0.14424403
#> [17,] -0.9384459016 -0.17337618  1.482492407 -0.98696723
#> [18,]  0.7655024053  1.34003105  0.216416842  0.86976671
#> [19,]  0.9717602823  0.16752142 -0.473906137  0.14236976
#> [20,] -0.7920171991 -0.36436944 -2.644739730  1.13566133
#> [21,]  0.0947270969 -0.46555156 -1.027242110  0.04972567
#> [22,]  0.0002326194 -1.16784718 -1.428182410 -0.57416325
#> [23,]  1.3468675714  1.24849240 -0.891415878  0.55825680
#> [24,]  0.2166780164 -1.06321583  0.181952909 -0.06131613
#> [25,]  1.2650705390 -0.05501372 -0.748018867  1.31347802
#> [26,] -0.1868875684  0.79271792  0.471212913  0.30902115
#> [27,]  0.3422128834 -0.01208605  0.988927593  0.54096059
#> [28,] -0.4953049915 -0.74053710 -0.521613092  0.46199029
#> [29,]  1.4370897655  0.28778230 -1.676870076 -0.94856983
#> [30,] -1.6901108801 -1.92335324 -0.991408155 -0.69354732
#> [31,]  1.2648432173 -1.74463462  1.708301296 -1.62975958
#> [32,]  0.6124553367  0.23590321 -0.417709515  0.39442446
#> [33,] -0.9974077525  0.43424448  0.135709736  0.84857345
#> [34,]  0.7817376081  0.85900587 -2.265468081  2.24565139
#> [35,] -0.1457885379  0.05839107  0.029174411 -0.69094992
#> [36,]  0.3081209770  1.58879065 -0.024202845 -0.53764278
#> [37,] -1.8273079709  0.42062266  0.602137138  1.63691919
#> [38,] -0.1584068920 -0.18242887  0.051802667  1.24247059
#> [39,] -0.0387130185  0.25704472 -0.518988435 -0.40271666
#> [40,]  0.0707100023 -0.11685602  0.009867921  0.94915882
#> [41,]  0.1367623317 -0.51247516  1.145732815 -0.55681221
#> [42,] -0.0454290541 -1.05699156  0.434901780  1.61457648
#> [43,] -0.0082067315  1.11147720  0.084712862  0.42268236
#> [44,] -0.1663995699 -0.16177723  0.341024643 -0.23477282
#> [45,]  1.1273127203  1.23172384 -1.007492261 -1.46585039
#> [46,]  0.3532137409 -1.75714140  0.059660386 -1.39362460
#> [47,] -0.8982903821 -3.36249538 -1.146582273 -0.85372947
#> [48,]  0.9503902621 -0.62434394 -0.133753236  1.23526576
#> [49,] -0.5085102560 -0.13337581  0.338078765 -1.49182553
#> [50,] -0.2093932759  0.43597670  0.941608091 -0.59260715
#> 
#> $datablocks[[2]]
#>              [,1]         [,2]        [,3]          [,4]
#>  [1,] -0.28289853 -1.705773384  0.89195904  1.1171111126
#>  [2,]  1.40643943 -0.841023748 -0.26128417  0.1789003171
#>  [3,] -1.05234632  0.258710352  0.60272604  0.2925219027
#>  [4,]  0.39170177  1.693140375 -0.26231545 -0.0006482874
#>  [5,] -1.96451520  0.729603906  0.33569445 -0.2221221594
#>  [6,]  0.24360695 -0.820930410 -0.81537008  1.0850571203
#>  [7,]  0.90673555 -0.155088057  1.56458781  0.6094595623
#>  [8,]  0.76065790 -2.214650158  0.08471075 -1.0021593115
#>  [9,]  1.87221100 -1.310021111 -0.13776775 -0.7747888718
#> [10,]  0.79015807 -0.347260075  0.28525401 -1.2103967371
#> [11,]  1.36607356  0.535233353  0.40328941 -1.3225062861
#> [12,]  0.03091517  0.185686191 -0.78581261 -1.6156892538
#> [13,] -0.38507446  0.683267635  0.26012980 -0.2589452500
#> [14,] -0.17132522  0.238875077  0.20059330 -0.4083668062
#> [15,] -0.04227477  2.673672962 -0.86321048  1.6471424361
#> [16,] -1.22254379 -1.148938079  1.17111274  1.4252892910
#> [17,]  0.07743379  0.229307969  0.06033041  0.0278596737
#> [18,] -1.31879978 -0.139748226  0.07254711  0.8769337932
#> [19,]  1.36437692 -0.033131469 -0.01223735  0.5924750334
#> [20,] -2.40958616  0.318772165 -0.14624094 -0.6608371212
#> [21,]  0.91571287 -1.286561564 -0.62216896  1.6257437452
#> [22,]  1.06798211 -1.581203994 -1.10588692  0.5804317838
#> [23,]  0.15312136  0.624525204  0.43861589  0.0363837903
#> [24,]  1.34024905  0.112263508  1.77110852  0.0731020470
#> [25,] -0.93576543  0.591525332  1.52976530  0.2872567199
#> [26,] -0.45651861 -0.722113613  0.43976085 -0.8682356991
#> [27,] -0.15234054 -0.134794813 -1.50553743  0.0984715224
#> [28,] -1.67972020 -0.348811881  1.52525580  0.7856693214
#> [29,]  0.40695668  0.300540770  1.48732013 -0.4655596630
#> [30,]  0.49270726  0.534524177  0.13428412  0.0021721434
#> [31,] -1.19401458  0.089295155 -3.18913654  0.0252996496
#> [32,]  0.68711386  0.003479698 -1.02803333  0.4696239734
#> [33,] -1.04657755 -0.065617915  0.43873635  0.1677858440
#> [34,] -0.46635773  1.985714613 -0.54964796 -1.8377295114
#> [35,] -0.35967630 -2.050964431 -1.26424380 -1.6053371406
#> [36,]  0.43670460  2.314797480  0.19210080  1.4810069698
#> [37,]  0.45178030 -0.329948590  0.70671070 -0.3537071511
#> [38,]  0.10981816  0.638220600  0.84764298  0.4466070867
#> [39,] -1.18595654  0.434679559  0.67616806 -0.8100418885
#> [40,] -1.36982243 -0.268121302 -0.50709963  1.4681619692
#> [41,]  1.75393615  0.080224246 -0.65728912 -1.3262997315
#> [42,] -0.05248658 -0.898349968 -0.30828379 -2.1525009542
#> [43,]  0.82602396 -0.442141862  0.84787805 -0.5632059206
#> [44,] -0.61124496  1.703605931 -1.65800328  0.4851152279
#> [45,]  0.36144955 -0.434661298  0.69787682  0.1609110767
#> [46,] -0.86163456  0.644827762  0.23123730 -1.5344668667
#> [47,]  0.66372236  0.217352482 -2.47544996 -0.4596144266
#> [48,]  0.34735667 -0.384272069 -0.11352976 -0.0185725736
#> [49,] -1.12937121 -0.187519582 -0.52339070  1.2287429118
#> [50,]  1.12590642  0.029801100  0.89454345  2.1964955863
#> 
#> $datablocks[[3]]
#>              [,1]         [,2]         [,3]        [,4]
#>  [1,] -1.13836429  0.494196638 -0.897201981 -0.67681181
#>  [2,]  1.16813448 -0.472052501  1.428269157  0.94657337
#>  [3,] -0.41773865 -1.579484677 -1.034421598  0.79903990
#>  [4,]  1.94810856  0.483423870 -0.923078905  0.76108519
#>  [5,] -1.28131715  0.982696596 -1.321664251  0.19988793
#>  [6,]  0.09415252 -0.288507938 -0.203641507 -0.78029347
#>  [7,]  0.04805402  0.418739661 -0.011261569  1.34994077
#>  [8,] -1.57430603 -0.351827099  1.387643440  0.16712425
#>  [9,]  0.79172877  0.286715530  1.403052467 -1.31759570
#> [10,]  0.87849017  1.300311128 -1.105248639 -0.33429635
#> [11,] -0.17645881 -0.895901634  0.740965184  0.32014221
#> [12,] -0.07027107 -0.144945988 -0.515385641 -0.83442471
#> [13,] -0.11158945 -1.411083873  0.059678543  0.23904919
#> [14,] -2.15585542  0.078240335 -0.434142248  0.67772371
#> [15,] -1.10035856  0.156686187  0.821075834 -2.01857547
#> [16,] -0.79407820  0.155801095  0.398028937 -1.04085053
#> [17,]  1.59727351  1.293798779  1.616186261  1.03977510
#> [18,]  0.31017686 -0.004068852 -1.185082527 -0.79355753
#> [19,] -0.18173864  0.679191290 -1.577414238  0.51724256
#> [20,] -1.04099270  2.494033783 -0.966708054  1.97339421
#> [21,] -0.91595864 -0.614702156  0.816640283  1.95645795
#> [22,]  0.98408889  1.432790280 -0.152010547 -0.28691835
#> [23,] -0.55755188  1.731342073  0.550601784  0.08097017
#> [24,] -0.53134093  0.163618820  0.867213841 -0.72032751
#> [25,]  1.15194518 -2.560192782  0.002928587  0.40561941
#> [26,]  1.23831901 -0.787391124  0.528850016 -0.77426863
#> [27,] -0.78220463 -0.758488088 -0.110692727  1.09177530
#> [28,]  0.07043723  0.936282009 -0.871132529  0.73465075
#> [29,]  0.03012184  0.021752162 -0.811253230 -1.24002597
#> [30,] -0.05356218 -0.220140997 -1.052915837 -1.77717615
#> [31,] -0.06695137 -1.431882830  1.002550613  1.58172628
#> [32,]  1.69265038 -1.414842404  0.047089098  1.16418815
#> [33,]  0.95334727  0.192703117  1.915888677 -1.39948775
#> [34,] -0.34014817 -0.671376504  1.005979272  0.32476291
#> [35,] -0.91493908  1.067599180  0.281437434 -0.69315182
#> [36,] -1.34044342  0.119219746  0.809166086 -1.50181057
#> [37,]  0.92327033 -1.730888993 -1.376070253 -0.19885508
#> [38,]  0.87717895  0.123448707 -2.084112254 -0.25794448
#> [39,]  1.35433021  0.738776005  0.097268803  0.44552453
#> [40,] -1.96144696  0.709019756  0.859866222 -0.73856330
#> [41,]  0.26980551 -1.648504320  0.994749426  0.66848050
#> [42,]  0.20461774 -0.079293417  0.150486489 -0.66521905
#> [43,] -1.55078782  1.227192001  0.700777440 -0.12541329
#> [44,] -0.62836711 -1.145602033  0.374128031 -1.03265936
#> [45,]  0.88481348  0.513131836 -0.430692940  1.02474967
#> [46,]  1.11935699  1.058322816 -0.813609405 -1.57589079
#> [47,] -0.32374491  0.152374412 -2.581157746 -0.10705422
#> [48,]  1.07154420 -0.394629006  0.747438851  0.53024130
#> [49,] -0.35541733  0.237226464  0.381010856  0.23301365
#> [50,]  0.70398730 -0.642827058  0.469926991  1.65803293
#> 
#> $datablocks[[4]]
#>               [,1]         [,2]          [,3]        [,4]
#>  [1,] -1.899360069  0.324553308 -1.3233035383  0.48666250
#>  [2,] -0.029630215 -0.748137571  1.2257388020  0.34075008
#>  [3,]  0.693937688 -0.323024042 -0.1173003762 -2.03491587
#>  [4,]  1.648126914 -0.075388746 -0.2519904843 -0.26258083
#>  [5,] -0.007819842 -0.721239978  0.6197446032 -0.19883780
#>  [6,]  0.146303883 -0.112807568  0.5135736214  1.96863721
#>  [7,]  1.993969493 -0.351689443 -0.0004307029  0.23491397
#>  [8,] -0.690045237 -0.240561717  0.0234113152 -0.88243020
#>  [9,] -0.517450639  0.721225165  0.3878275082 -0.49203306
#> [10,]  2.102886247 -0.215287230 -0.1455444647  1.35819620
#> [11,] -0.056829890  1.592648550  1.0365435483 -1.03678455
#> [12,] -0.317732247 -0.783576280  0.6306958184 -0.13916010
#> [13,] -1.016998059  0.004621279 -0.8872231847  0.54725086
#> [14,]  0.014218315  0.480307252 -0.9596198929  0.47668594
#> [15,]  0.119997714  0.064028031  0.7129686455 -1.15431840
#> [16,] -0.111653623  0.206449098  0.5682558936  0.02861876
#> [17,] -0.468627179 -0.252423458  0.2726780539  1.32754393
#> [18,] -1.808560095  0.780702029  0.3668998489 -2.26977810
#> [19,] -0.844838697  1.179403352 -0.4337400571 -0.47930334
#> [20,] -0.843486030 -0.187045710 -0.7456808430 -2.66401507
#> [21,]  1.422010406 -2.094739901  1.5010913264  2.39452364
#> [22,]  2.008230860  2.177833330  0.5435903336  0.48925204
#> [23,]  0.602498248  0.318072943  0.3598774751 -0.58975949
#> [24,]  0.547334848 -2.323114489  0.4497906209 -0.02351194
#> [25,]  1.007637738  0.411740004 -0.3895632946  0.74616255
#> [26,]  0.693481564 -0.299812274  0.0491028426 -0.12224401
#> [27,]  0.158800328  0.399260054 -2.3828081032 -0.12958622
#> [28,] -0.717746501 -0.036130271 -1.7926489424  1.87335979
#> [29,] -1.179176971 -1.356620073  1.2800925231 -0.59504054
#> [30,] -0.374252393 -1.101830823 -1.8842957436 -1.09309879
#> [31,]  0.406946901 -1.252595777  0.4384654064  0.74274888
#> [32,]  1.689547785  0.065214075  0.4478244968  0.04277510
#> [33,] -0.604479257  1.264506407  0.3928927544 -0.27042361
#> [34,]  1.529026705  0.803340241 -1.5646410757  0.29258484
#> [35,] -2.106684878 -0.970547888  0.2931400040  0.99294122
#> [36,] -1.352718626  1.076000857  1.0573650213  0.54309748
#> [37,] -0.869334176 -0.160569744  0.2822390121  0.23578543
#> [38,]  0.316821433  0.863990056 -1.5947086702 -0.08877839
#> [39,]  0.348386456  0.532334634 -1.9430869338 -1.42229969
#> [40,]  0.049295629  0.173441078 -0.8358688494 -0.12216444
#> [41,] -0.594103042 -0.734815141  1.5637128523  0.34665331
#> [42,]  1.224457921 -1.857021730  0.0282808958  0.64253442
#> [43,] -0.651308930  1.374505000  0.6436829310  0.43876687
#> [44,] -0.306358857  0.089839120  0.7572900449  0.49430521
#> [45,] -0.213342039 -0.534561558 -1.2686326258 -0.50000394
#> [46,]  0.187256274  1.079648415  1.6746473689  0.48354709
#> [47,] -0.428717992 -1.057850148 -0.4979495578  0.19192542
#> [48,]  0.420781473 -0.285402165 -0.5536094854 -0.64828575
#> [49,] -0.593642789 -0.624053842 -0.2788441240  0.82361177
#> [50,] -0.727056547  2.717183289  1.7300673817 -1.32448037
#> 
#> $datablocks[[5]]
#>              [,1]        [,2]       [,3]        [,4]
#>  [1,]  0.48999336 -0.85215206  0.3436153 -0.03547033
#>  [2,]  0.13640515  0.53020698  0.0799988 -1.27538967
#>  [3,] -0.16268949 -0.91149757 -0.3132136 -0.24528957
#>  [4,] -1.99914169  0.62278678 -1.6656853 -0.15778884
#>  [5,]  1.05061004  0.90055896 -0.5255358  1.31032869
#>  [6,] -0.15515085 -1.37399184 -0.9406324 -0.63515790
#>  [7,]  1.21028711  0.58875716  1.1516262 -0.82005469
#>  [8,]  0.25908801  1.23667747 -0.1810264 -0.69651585
#>  [9,] -1.59974168 -0.12948896  1.9572143 -0.34780143
#> [10,]  0.64954806  0.03756448 -0.5441540 -0.22273174
#> [11,] -0.46600084  0.39061114  0.2576567  0.43592450
#> [12,] -1.00881633  1.26348471  0.5106784  1.01735179
#> [13,] -1.01915690 -0.21431853 -0.2766500 -0.50229001
#> [14,] -0.11911533  1.24663196 -1.2979772  0.42211289
#> [15,] -0.73870983  0.34829269  0.9191967 -1.19407883
#> [16,] -0.72642276 -0.86803905  0.1610629 -0.05194628
#> [17,]  0.60634373  0.51948644 -0.8070434 -0.71352036
#> [18,]  0.76474008  1.35016295  0.8365227  2.25223389
#> [19,] -0.78835039  1.32389462  0.9756023  0.50815625
#> [20,] -0.08178376  0.95393259  1.8571479 -1.24949075
#> [21,] -1.11369002 -1.36802005  0.3548181  2.71716947
#> [22,] -0.27889329 -0.09303306 -0.1739618 -0.87541475
#> [23,]  1.05916310 -0.06313504 -1.3786497 -0.78468436
#> [24,]  2.15148730 -0.79554761 -0.6247984  0.36373648
#> [25,]  1.80261070 -1.27801811 -1.4369180  0.24795033
#> [26,]  1.33604459 -0.07406958  0.8157789 -0.54869980
#> [27,] -1.08823582 -0.29653397 -0.5572153 -1.11022401
#> [28,] -1.09775109  0.47170851 -0.8093033 -0.74450795
#> [29,] -0.50285604 -2.18143881  0.3931265  0.27210757
#> [30,]  1.36000282 -0.02988600  1.3662946  0.21309812
#> [31,] -0.15392750  1.36196958 -0.6577308  0.75233064
#> [32,]  0.15615301  1.53860394 -0.5965790 -0.68462162
#> [33,] -0.93233177 -0.53052032  0.5665541  0.75951543
#> [34,] -0.86153244 -1.19468744 -0.4579725  0.59494178
#> [35,] -1.22287974 -1.67279544  1.7710420 -1.13050201
#> [36,]  1.06956320  0.85860856 -1.2475747 -1.39664433
#> [37,]  0.31503194 -0.58482291 -0.1565045  0.50014935
#> [38,]  0.92222896 -1.29497115  1.1306381 -1.22596693
#> [39,]  1.24532020 -0.77144960 -0.1759539  1.67922491
#> [40,]  0.29206918  1.78348331 -0.9330501  1.57850960
#> [41,]  0.05446038 -0.36075652 -0.5919588 -1.11188974
#> [42,] -1.78886668 -1.14466801  0.3569504  0.11031837
#> [43,] -0.18737784 -0.78505870  0.6639892  1.33603749
#> [44,] -0.55865024 -0.27816030  0.7125571  1.16835587
#> [45,] -0.63251409  0.69474871 -0.3626306 -0.09932113
#> [46,]  0.04029759 -0.07619960 -0.1763163 -0.15557402
#> [47,]  1.95929549  0.77389469  0.3787596  0.53277669
#> [48,]  0.76264849 -0.66865201 -1.9609993 -1.62705479
#> [49,] -1.25317004 -0.94295973 -1.3371944  1.48056063
#> [50,]  0.84436396  2.03880575  2.6263985 -0.61025906
#> 
#> 
#> $indices
#> $indices[[1]]
#> [1]  4 13 21 46 50
#> 
#> $indices[[2]]
#> [1]  5  7 18 20 26
#> 
#> $indices[[3]]
#> [1] 10 28 32 33 44
#> 
#> $indices[[4]]
#> [1] 12 17 27 31 34
#> 
#> $indices[[5]]
#> [1]  8 16 29 38 42
#> 
#> $indices[[6]]
#> [1]  1  3  9 23 24
#> 
#> $indices[[7]]
#> [1] 25 30 37 40 41
#> 
#> $indices[[8]]
#> [1]  6 11 14 45 48
#> 
#> $indices[[9]]
#> [1]  2 15 22 35 49
#> 
#> $indices[[10]]
#> [1] 19 36 39 43 47
#> 
#> 
#> $z
#>  [1] 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5
#> 
#> $pars
#> $pars$family
#> [1] "gaussian"
#> 
#> $pars$mu.x
#>  [1]  0.06197652 -0.21460240 -0.09834921 -0.05568752  0.10072770 -0.19108954
#>  [7] -0.03025647 -0.03311304  0.11105047 -0.01639079 -0.24558517 -0.04003557
#> [13] -0.13551608  0.12815370  0.23468486 -0.06882579  0.32864523  0.04517092
#> [19]  0.02261847 -0.04722240
#> 
#> $pars$sd.x
#>  [1] 0.9643406 1.0529562 0.9523880 1.0805635 0.9842422 0.9643788 0.8338837
#>  [8] 0.9307574 0.9543343 0.8125298 0.9872985 0.9622111 1.0113876 0.9773881
#> [15] 1.0793100 0.9760331 0.9922422 0.9266769 0.9137116 1.2479439
#> 
#> $pars$mu.y
#> [1] 0.1209037
#> 
#> $pars$sd.y
#> [1] 0.8294677
#> 
#> 
#> attr(,"class")
#> [1] "multiridge"

# \donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x %*% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
#>     glmnet multiridge 
#>  0.4692074  0.5982584 
sapply(coef, function(x) mean((beta-x[-1])^2))
#>     glmnet multiridge 
#>  0.1718059  0.1262343 

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
#>     glmnet multiridge 
#>   73.65774   50.10826 
# }