Multi-penalty ridge regression

Fits multi-penalty ridge regression (tuning regularisation hyperparameters and estimating regression coefficients). This is a wrapper function of some functions from the multiridge-package.

Usage

multiridge(
  x,
  y,
  z,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10,
  penalties = NULL
)

Arguments

x

predictors: \(n \times p\) matrix

y

response: \(n\)-dimensional vector

z

\(p\)-dimensional vector with entries in \(\{1, \ldots, q\}\)

family

character "linear" (or "gaussian"), "logistic" (or "binomial"), or "cox"

foldid

\(n_0\)-dimensional vector containing the fold identifiers

nfolds

integer specifying the number of folds

penalties

\(q\)-dimensional vector of penalty parameters, or NULL (cross-validation)

Value

Returns an object of class "multiridge", a list with the following slots:

slots from IWLSridge() or IWLSCoxridge()

character family with value "gaussian" (also for "linear"), "binomial" (also for "logistic"), or "cox"

\(q\)-dimensional vector penalties containing optimised regularisation hyperparameters (one for each predictor group)

list indices with nfolds slots (one for each cross-validation fold), each containing the indices of the observations

list datablocks with \(q\) slots (one for each predictor group), each containing an \(n_0 \times p_k\) matrix, where \(k \in \{1, \ldots, q\}\)

\(p\)-dimensional group vector z (see argument)

list pars with slots family (see above), the \(p\)-dimensional vectors mu.x and sd.x and the scalars mu.y and sd.y

Details

The number of observations (samples) for training or testing are indicated by \(n_0\) and \(n_1\), respectively, the number of variables (features) is indicated by \(p\), and the number of variable groups is indicated by \(q\). Observations (samples) are indexed by \(i\) in \(\{1, \ldots, n\}\), variables (features) are indexed by \(j\) in \(\{1, \ldots, p\}\), and variable groups are indexed by \(k\) in \(\{1, \ldots, q\}\). The number of variables in the \(k^{\text{th}}\) group is indicated by \(p_k\), with \(\sum_{k=1}^q p_k = p\) for non-overlapping groups.

References

Mark A. van de Wiel, Mirrelijn M. van Nee and Armin Rauschenberger (2021). "Fast cross-validation for multi-penalty high-dimensional ridge regression" Journal of Computational and Graphical Statistics 30(4):835-847. doi: 10.1080/10618600.2021.1904962. # nolint: line_length_linter.

See also

Examples

# minimal example
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
multiridge(x = x, y = y, z = z)
#> $etas
#>               [,1]
#>  [1,] -0.084733142
#>  [2,]  0.079847419
#>  [3,]  0.066148869
#>  [4,]  0.045355900
#>  [5,] -0.239275208
#>  [6,]  0.115652923
#>  [7,]  0.016734376
#>  [8,]  0.067006160
#>  [9,]  0.001489244
#> [10,] -0.119032904
#> [11,]  0.057466620
#> [12,] -0.160205609
#> [13,]  0.014289111
#> [14,] -0.166391190
#> [15,]  0.220483121
#> [16,]  0.837425738
#> [17,] -0.079442133
#> [18,]  0.097014529
#> [19,] -0.163440587
#> [20,]  0.089442384
#> [21,] -0.132265901
#> [22,]  0.056419746
#> [23,] -0.253826307
#> [24,]  0.221430474
#> [25,]  0.141116337
#> [26,] -0.430003236
#> [27,]  0.135496013
#> [28,] -0.151737815
#> [29,] -0.143557653
#> [30,] -0.350995726
#> [31,]  0.126988415
#> [32,] -0.053493346
#> [33,] -0.029755455
#> [34,] -0.226362528
#> [35,]  0.141945566
#> [36,] -0.063463587
#> [37,]  0.238723118
#> [38,] -0.264592864
#> [39,]  0.451045397
#> [40,]  0.019530949
#> [41,]  0.139142838
#> [42,] -0.182700230
#> [43,] -0.558736968
#> [44,] -0.160472163
#> [45,]  0.112939453
#> [46,]  0.266466526
#> [47,]  0.081948088
#> [48,] -0.337031851
#> [49,] -0.187314871
#> [50,]  0.697281961
#> 
#> $Ypred
#>               [,1]
#>  [1,] -0.084733142
#>  [2,]  0.079847419
#>  [3,]  0.066148869
#>  [4,]  0.045355900
#>  [5,] -0.239275208
#>  [6,]  0.115652923
#>  [7,]  0.016734376
#>  [8,]  0.067006160
#>  [9,]  0.001489244
#> [10,] -0.119032904
#> [11,]  0.057466620
#> [12,] -0.160205609
#> [13,]  0.014289111
#> [14,] -0.166391190
#> [15,]  0.220483121
#> [16,]  0.837425738
#> [17,] -0.079442133
#> [18,]  0.097014529
#> [19,] -0.163440587
#> [20,]  0.089442384
#> [21,] -0.132265901
#> [22,]  0.056419746
#> [23,] -0.253826307
#> [24,]  0.221430474
#> [25,]  0.141116337
#> [26,] -0.430003236
#> [27,]  0.135496013
#> [28,] -0.151737815
#> [29,] -0.143557653
#> [30,] -0.350995726
#> [31,]  0.126988415
#> [32,] -0.053493346
#> [33,] -0.029755455
#> [34,] -0.226362528
#> [35,]  0.141945566
#> [36,] -0.063463587
#> [37,]  0.238723118
#> [38,] -0.264592864
#> [39,]  0.451045397
#> [40,]  0.019530949
#> [41,]  0.139142838
#> [42,] -0.182700230
#> [43,] -0.558736968
#> [44,] -0.160472163
#> [45,]  0.112939453
#> [46,]  0.266466526
#> [47,]  0.081948088
#> [48,] -0.337031851
#> [49,] -0.187314871
#> [50,]  0.697281961
#> 
#> $convergence
#> [1] FALSE
#> 
#> $nIt
#> [1] 1
#> 
#> $Hresl
#> $Hresl$Hmatl
#>               [,1]
#>  [1,] -0.084733142
#>  [2,]  0.079847419
#>  [3,]  0.066148869
#>  [4,]  0.045355900
#>  [5,] -0.239275208
#>  [6,]  0.115652923
#>  [7,]  0.016734376
#>  [8,]  0.067006160
#>  [9,]  0.001489244
#> [10,] -0.119032904
#> [11,]  0.057466620
#> [12,] -0.160205609
#> [13,]  0.014289111
#> [14,] -0.166391190
#> [15,]  0.220483121
#> [16,]  0.837425738
#> [17,] -0.079442133
#> [18,]  0.097014529
#> [19,] -0.163440587
#> [20,]  0.089442384
#> [21,] -0.132265901
#> [22,]  0.056419746
#> [23,] -0.253826307
#> [24,]  0.221430474
#> [25,]  0.141116337
#> [26,] -0.430003236
#> [27,]  0.135496013
#> [28,] -0.151737815
#> [29,] -0.143557653
#> [30,] -0.350995726
#> [31,]  0.126988415
#> [32,] -0.053493346
#> [33,] -0.029755455
#> [34,] -0.226362528
#> [35,]  0.141945566
#> [36,] -0.063463587
#> [37,]  0.238723118
#> [38,] -0.264592864
#> [39,]  0.451045397
#> [40,]  0.019530949
#> [41,]  0.139142838
#> [42,] -0.182700230
#> [43,] -0.558736968
#> [44,] -0.160472163
#> [45,]  0.112939453
#> [46,]  0.266466526
#> [47,]  0.081948088
#> [48,] -0.337031851
#> [49,] -0.187314871
#> [50,]  0.697281961
#> 
#> $Hresl$Mmatl
#>              [,1]
#>  [1,] -0.85125196
#>  [2,]  0.67898438
#>  [3,]  0.49954185
#>  [4,] -1.50461933
#>  [5,] -1.53232873
#>  [6,]  0.33879086
#>  [7,] -1.58803396
#>  [8,]  1.77118246
#>  [9,]  1.71850144
#> [10,] -1.31838745
#> [11,] -0.06052245
#> [12,]  1.59563479
#> [13,]  1.22435151
#> [14,] -0.08269430
#> [15,] -0.53558614
#> [16,] -0.34469101
#> [17,] -1.96256795
#> [18,]  1.50802476
#> [19,] -0.68527376
#> [20,]  0.43198673
#> [21,]  0.88206435
#> [22,]  0.51012233
#> [23,] -0.25033756
#> [24,]  0.28410349
#> [25,] -0.74347604
#> [26,]  1.12073174
#> [27,]  0.75158926
#> [28,]  1.29120431
#> [29,] -0.17641763
#> [30,] -0.27066302
#> [31,] -0.48779544
#> [32,] -0.07935285
#> [33,] -0.73129976
#> [34,] -1.33222160
#> [35,]  0.01093272
#> [36,]  0.35969715
#> [37,]  1.38471691
#> [38,] -1.13311200
#> [39,] -0.36144451
#> [40,] -0.61363919
#> [41,]  0.18158147
#> [42,] -0.07513299
#> [43,] -0.39786463
#> [44,] -0.71712829
#> [45,] -0.68639095
#> [46,]  0.25463760
#> [47,] -0.27131216
#> [48,] -0.08353130
#> [49,]  1.25659342
#> [50,]  0.82210340
#> 
#> 
#> $linearized
#>               [,1]
#>  [1,] -0.935985099
#>  [2,]  0.758831799
#>  [3,]  0.565690716
#>  [4,] -1.459263427
#>  [5,] -1.771603936
#>  [6,]  0.454443787
#>  [7,] -1.571299584
#>  [8,]  1.838188617
#>  [9,]  1.719990688
#> [10,] -1.437420358
#> [11,] -0.003055826
#> [12,]  1.435429176
#> [13,]  1.238640625
#> [14,] -0.249085492
#> [15,] -0.315103024
#> [16,]  0.492734729
#> [17,] -2.042010088
#> [18,]  1.605039292
#> [19,] -0.848714344
#> [20,]  0.521429111
#> [21,]  0.749798453
#> [22,]  0.566542080
#> [23,] -0.504163868
#> [24,]  0.505533966
#> [25,] -0.602359702
#> [26,]  0.690728500
#> [27,]  0.887085275
#> [28,]  1.139466493
#> [29,] -0.319975283
#> [30,] -0.621658742
#> [31,] -0.360807023
#> [32,] -0.132846193
#> [33,] -0.761055216
#> [34,] -1.558584133
#> [35,]  0.152878288
#> [36,]  0.296233558
#> [37,]  1.623440032
#> [38,] -1.397704862
#> [39,]  0.089600890
#> [40,] -0.594108242
#> [41,]  0.320724307
#> [42,] -0.257833219
#> [43,] -0.956601594
#> [44,] -0.877600455
#> [45,] -0.573451498
#> [46,]  0.521104126
#> [47,] -0.189364068
#> [48,] -0.420563147
#> [49,]  1.069278550
#> [50,]  1.519385365
#> 
#> $unpen
#> [1] FALSE
#> 
#> $intercept
#> [1] TRUE
#> 
#> $eta0
#> [1] 0
#> 
#> $X1
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1
#> [31,]    1
#> [32,]    1
#> [33,]    1
#> [34,]    1
#> [35,]    1
#> [36,]    1
#> [37,]    1
#> [38,]    1
#> [39,]    1
#> [40,]    1
#> [41,]    1
#> [42,]    1
#> [43,]    1
#> [44,]    1
#> [45,]    1
#> [46,]    1
#> [47,]    1
#> [48,]    1
#> [49,]    1
#> [50,]    1
#> 
#> $family
#> [1] "gaussian"
#> 
#> $penalties
#> [1] 1.000000e+06 1.899254e+01 1.000000e+06 2.156563e+03 1.000000e+06
#> 
#> $datablocks
#> $datablocks[[1]]
#>              [,1]        [,2]        [,3]         [,4]
#>  [1,] -1.14608439  0.33024733  1.37405760 -0.405509646
#>  [2,] -0.60584593 -0.82740913 -0.26623829  0.824110134
#>  [3,]  0.12253214 -0.75211060 -0.79275601 -0.258980153
#>  [4,]  0.73546908 -1.13642102 -1.17177558  2.097448632
#>  [5,]  0.17992705  1.30170347  0.83499204 -1.093717318
#>  [6,]  0.76667950  0.49922280 -0.30928661  0.059720317
#>  [7,]  0.97637154  1.14500633 -0.87225581  1.420438075
#>  [8,]  0.39905242  1.48862957 -0.45723470  0.039159111
#>  [9,]  0.97185318  0.47480079  1.56818544 -0.761213959
#> [10,] -0.78377080  0.22209228 -0.13165823 -0.400625305
#> [11,]  0.07542905 -1.73837563 -0.75310317  0.602525102
#> [12,] -0.64789603 -0.03511441 -0.46802330  0.927114738
#> [13,] -1.05169580 -0.42998066 -0.06618017  0.978458924
#> [14,]  2.15517058 -0.50853845  0.70509233 -0.158188352
#> [15,]  0.56470254  0.20702946  0.09392394  0.038776053
#> [16,]  0.75119907  2.24797699  0.76783522  0.625581005
#> [17,] -2.22530808  0.97528476 -0.37574203 -1.847922465
#> [18,] -0.86075881 -1.16863622 -0.59319916 -1.188459804
#> [19,]  0.69343051  1.06219313  1.77059201  3.001698678
#> [20,]  1.48415282  0.45465223  1.37717606 -0.508497671
#> [21,] -0.32540527  1.15798974  1.41926953  0.792177746
#> [22,]  0.48566855  1.75587080 -0.13782668 -2.194685549
#> [23,]  1.02142564 -0.96581461 -0.14626706  0.229925455
#> [24,]  1.11712458  0.50012629 -1.13166817 -1.113746532
#> [25,]  0.57830331 -0.31757993  2.00462754 -0.005731694
#> [26,] -0.15106315 -0.34311015 -0.09830307  0.011725354
#> [27,] -1.69218265  1.15790345 -1.09172947 -0.379269081
#> [28,] -0.81469541 -0.20543279  0.32419378  0.576975345
#> [29,] -2.40491250  0.20033937  0.70124795  1.080520779
#> [30,]  0.65392291 -0.10065744 -0.52969374 -0.858691929
#> [31,] -0.25470319 -0.69564965 -0.03297746  0.452377439
#> [32,] -2.23545677 -0.43740711  1.15899175 -0.711419415
#> [33,] -0.61778538  0.19938622  0.03243857  0.590525062
#> [34,] -0.13450458  1.55630531  1.04457669 -1.912197499
#> [35,] -0.11146826 -0.97067417  0.44537018  0.065786352
#> [36,] -0.86792858 -0.33183239  1.35874623 -1.015656245
#> [37,]  0.61155738  1.24526618  0.13918272  0.759389001
#> [38,]  0.79064476  0.02341180 -0.52577974  0.063994536
#> [39,] -0.74078904 -0.53086355 -2.61686303  1.013585342
#> [40,]  0.02914379 -0.63630390 -1.05878771 -0.024573591
#> [41,] -0.05290286 -1.36815544 -1.44499859 -0.621013949
#> [42,]  1.11633889  1.14987515 -0.92795135  0.461584251
#> [43,]  0.13503003 -1.25912070  0.10598489 -0.130729999
#> [44,]  1.04531704 -0.20848868 -0.78982234  1.183578828
#> [45,] -0.21537357  0.67491952  0.38461830  0.223313947
#> [46,]  0.24402809 -0.16375441  0.88331360  0.445049006
#> [47,] -0.48316295 -0.92286213 -0.57173408  0.369553078
#> [48,]  1.19467605  0.14873382 -1.68455018 -0.978948168
#> [49,] -1.52057619 -2.15545677 -1.02427019 -0.735145610
#> [50,]  1.04511967 -1.96921683  1.57625954 -1.630168355
#> 
#> $datablocks[[2]]
#>              [,1]         [,2]        [,3]        [,4]
#>  [1,]  0.77589266  0.218077043 -0.63067555  0.04326558
#>  [2,] -0.79219139  0.424317142 -0.03833006  0.50486185
#>  [3,]  0.94078173  0.865994419 -2.60840121  1.92484951
#>  [4,]  0.03732667  0.033495579 -0.15235883 -1.05990710
#>  [5,]  0.47945636  1.624842462 -0.20949052 -0.90408598
#>  [6,] -1.60055410  0.410152841  0.46090478  1.30613513
#>  [7,]  0.02503579 -0.216914875 -0.12813896  0.90521824
#>  [8,]  0.14162337  0.240060500 -0.73907819 -0.76694730
#>  [9,]  0.24820666 -0.148730622 -0.17302330  0.60709665
#> [10,]  0.31254480 -0.560105079  1.04273573 -0.92356976
#> [11,]  0.13508163 -1.126306530  0.28190647  1.28342603
#> [12,]  0.17133796  1.128520276 -0.09291393  0.07198686
#> [13,]  0.01725053 -0.195440803  0.18142623 -0.59624946
#> [14,]  1.27738851  1.253555672 -1.26194226 -1.84751463
#> [15,]  0.52337888 -1.854339444 -0.11972854 -1.77410446
#> [16,] -0.69564631 -3.523625739 -1.41081554 -1.22535596
#> [17,]  1.10505753 -0.676428970 -0.32674643  0.89789526
#> [18,] -0.31598172 -0.165908310  0.17827315 -1.87391572
#> [19,] -0.02462740  0.426118344  0.82425334 -0.95995167
#> [20,] -0.40090974 -1.387500983  0.96969091  1.21056834
#> [21,]  1.39579542 -0.609987009 -0.23330346  0.24113596
#> [22,] -1.21926038  0.378806275  0.66798033  0.35853862
#> [23,]  0.31656531  1.668531466 -0.23437923  0.05561237
#> [24,] -2.18940286  0.802196242  0.38942894 -0.17323166
#> [25,]  0.15905822 -0.591920680 -0.81129276  1.17744767
#> [26,]  0.86433247  0.006751714  1.67133703  0.68602327
#> [27,]  0.70897074 -1.845042435  0.12761775 -0.97922673
#> [28,]  1.89116946 -1.031672070 -0.10445851 -0.74428991
#> [29,]  0.74034582 -0.166034054  0.33681249 -1.19439384
#> [30,]  1.35286406  0.627433726  0.45993997 -1.31023415
#> [31,] -0.06715136  0.313148768 -0.78046017 -1.61317360
#> [32,] -0.50957953  0.760534358  0.31060442 -0.21128015
#> [33,] -0.28224530  0.360971976  0.24849950 -0.36567412
#> [34,] -0.14499294  2.550148246 -0.86119701  1.75823794
#> [35,] -1.40027488 -0.886839042  1.26088736  1.52900201
#> [36,] -0.01767622  0.352369995  0.10218564  0.08506902
#> [37,] -1.50264850  0.020544068  0.11492937  0.96239845
#> [38,]  1.35105959  0.116405362  0.02648729  0.66847353
#> [39,] -2.66276066  0.432809072 -0.11329724 -0.62654609
#> [40,]  0.87388035 -1.010579115 -0.60975690  1.73612715
#> [41,]  1.03582717 -1.275498113 -1.11434256  0.65602951
#> [42,]  0.06282174  0.707717845  0.49679044  0.09387680
#> [43,]  1.32539826  0.247132941  1.88676709  0.13181696
#> [44,] -1.09527010  0.678046988  1.63501228  0.35309822
#> [45,] -0.58556437 -0.503072489  0.49798479 -0.84084644
#> [46,] -0.26205403  0.024997783 -1.53123408  0.15803067
#> [47,] -1.88650750 -0.167429313  1.63030823  0.86809690
#> [48,]  0.33278976  0.416416855  1.59073601 -0.42477023
#> [49,]  0.42399029  0.626796092  0.17932973  0.05852666
#> [50,] -1.36993247  0.226481628 -3.28746402  0.08242380
#> 
#> $datablocks[[3]]
#>               [,1]         [,2]        [,3]        [,4]
#>  [1,]  0.776898039 -0.080647564 -0.89614141  0.61286848
#>  [2,] -1.225041959 -0.142779281  0.62714368  0.33084210
#>  [3,] -0.555046562  1.701753492 -0.39932360 -1.54303727
#>  [4,] -0.431858652 -1.927978249 -1.14145321 -1.32589843
#>  [5,]  0.487743747  1.997660724  0.37100513  1.55786735
#>  [6,]  0.505152060 -0.380462146  0.90544322 -0.15642167
#>  [7,]  0.110279200  0.490103584  1.05180570  0.59136236
#>  [8,] -1.385986623  0.307082004  0.87372376 -0.58280397
#>  [9,] -1.598301495 -0.324867816 -0.35513586  1.54586545
#> [10,]  2.008786362 -0.011639818 -0.51111224 -1.06517620
#> [11,] -0.077138442 -0.891561617 -0.14865954 -1.83714806
#> [12,]  0.937301257 -0.481344945  1.05204982 -0.35216957
#> [13,] -0.722351739  1.448084863 -1.55038444  0.62734294
#> [14,]  0.400844713 -0.474618515  0.89626897  0.32441857
#> [15,] -1.011483327  0.496044662  0.41164958 -1.25967987
#> [16,]  0.749887243  0.111664192 -2.39932778 -0.25537751
#> [17,]  0.384571295 -0.429309144  0.05359846  0.15671569
#> [18,] -1.320646023 -0.252391748 -0.37205464  1.32216114
#> [19,]  1.283583577 -0.056979686  1.10051327  2.22639345
#> [20,] -1.077992830  0.409142659 -0.91094362 -0.68347061
#> [21,]  1.331003050 -0.560631225  1.62755830  0.73313145
#> [22,] -0.325343740 -1.672103163 -1.06073360  0.60439047
#> [23,]  2.145637928  0.398330593 -0.93919107  0.57127034
#> [24,] -1.227298290  0.899424599 -1.37428985  0.08155711
#> [25,]  0.209295046 -0.376417116 -0.15384775 -0.77377101
#> [26,]  0.161148009  0.333410426  0.05615565  1.08511883
#> [27,] -1.533307127 -0.439967257  1.58321096  0.05296679
#> [28,]  0.937870474  0.200904688  1.60003157 -1.24263291
#> [29,]  1.028487413  1.218197745 -1.13804893 -0.38458399
#> [30,] -0.073341819 -0.986026502  0.87729207  0.18649367
#> [31,]  0.037564738 -0.232331472 -0.49415001 -0.82100715
#> [32,] -0.005589764 -1.503088055  0.13359441  0.11573010
#> [33,] -2.140699640 -0.008330996 -0.40546411  0.49852726
#> [34,] -1.038298185  0.070401017  0.96474146 -1.85432348
#> [35,] -0.718407219  0.069512695  0.50294026 -1.00113891
#> [36,]  1.779212326  1.211661640  1.83269003  0.81446136
#> [37,]  0.434919169 -0.090940435 -1.22519621 -0.78534553
#> [38,] -0.078856270  0.594812147 -1.65346848  0.35848783
#> [39,] -0.976294211  2.416274938 -0.98681698  1.62915819
#> [40,] -0.845703835 -0.703801246  0.95989958  1.61437923
#> [41,]  1.138778768  1.351160163 -0.09748700 -0.34324090
#> [42,] -0.471370056  1.650801033  0.66948994 -0.02221318
#> [43,] -0.443994325  0.077358938  1.01510606 -0.72144339
#> [44,]  1.314094325 -2.656388757  0.07164573  0.26108298
#> [45,]  1.404306466 -0.877120161  0.64574555 -0.76851361
#> [46,] -0.706006014 -0.848111690 -0.05238416  0.85983790
#> [47,]  0.184525887  0.852840697 -0.88248601  0.54820306
#> [48,]  0.142418950 -0.065025230 -0.81712131 -1.17494383
#> [49,]  0.055016146 -0.307800782 -1.08092205 -1.64367305
#> [50,]  0.041031958 -1.523962884  1.16284071  1.28738003
#> 
#> $datablocks[[4]]
#>              [,1]        [,2]        [,3]        [,4]
#>  [1,]  1.38098509 -1.48757673 -0.14146632  1.17366664
#>  [2,]  0.72281616  0.21948261  1.75932460 -1.23936577
#>  [3,] -0.42872573 -0.69808716  0.83383877  0.38356664
#>  [4,] -0.94043671  1.14853842  0.09689375 -0.57453466
#>  [5,] -1.31924445  0.14145031  0.63365641 -1.33567602
#>  [6,]  0.69603999 -1.82318793 -1.58898825 -0.10928317
#>  [7,]  0.65500687  0.14594107 -2.30915097 -0.16490042
#>  [8,]  1.07979357  0.79935971 -0.09042760  0.49723222
#>  [9,] -1.87209649  0.76776142  0.68522446 -0.61727773
#> [10,]  0.11428910 -1.73570330  0.82241668  0.70708713
#> [11,]  0.05625531 -0.06935165 -0.03629890 -0.54824324
#> [12,] -1.50650479  1.31801070  0.52341229 -0.04015687
#> [13,] -0.68531434 -1.20166925  0.19117108 -0.89409249
#> [14,]  0.66180355  0.55974722 -0.62742734  1.04242166
#> [15,]  0.87060729  1.13868781 -1.01689882 -1.40540324
#> [16,] -0.41412264  0.17665739 -2.81470522 -0.02287659
#> [17,]  0.82804167 -0.40420783  0.57087248  0.57697097
#> [18,] -0.44231920  0.26676214  0.19817173  0.29720863
#> [19,]  0.50082199 -0.66777039  0.28860999  1.63849272
#> [20,] -2.00451387  0.53270161 -1.22490774  0.50663717
#> [21,] -0.11595689 -0.63524752  1.28061352  0.37122019
#> [22,]  0.61489694 -0.17238260 -0.03949528 -1.83356464
#> [23,]  1.57869431  0.09724348 -0.17188575 -0.18871327
#> [24,] -0.09392690 -0.60596133  0.68496576 -0.12955525
#> [25,]  0.06174877  0.05650175  0.58060749  1.88201377
#> [26,]  1.92801933 -0.20359357  0.07537902  0.27299685
#> [27,] -0.78302192 -0.08259736  0.09881397 -0.76397700
#> [28,] -0.60868939  0.96459915  0.45700830 -0.40166104
#> [29,]  2.03803281 -0.05507843 -0.06725714  1.31548164
#> [30,] -0.14343045  1.91340762  1.09464849 -0.90722868
#> [31,] -0.40695994 -0.67383329  0.69573000 -0.07417015
#> [32,] -1.11326697  0.18435864 -0.79627281  0.56286741
#> [33,] -0.07166683  0.70228699 -0.86743345  0.49737821
#> [34,]  0.03517771  0.24904089  0.77659814 -1.01630835
#> [35,] -0.19880622  0.40410939  0.63435614  0.08154041
#> [36,] -0.55937430 -0.09551246  0.34382485  1.28703418
#> [37,] -1.91279959  1.02935775  0.43643795 -2.05153329
#> [38,] -0.93937406  1.46346498 -0.35053223 -0.38984697
#> [39,] -0.93800777 -0.02432897 -0.65714709 -2.41741289
#> [40,]  1.35030094 -2.10143228  1.55126481  2.27726629
#> [41,]  1.94242430  2.55055861  0.61011171  0.50904044
#> [42,]  0.52253675  0.52564577  0.42953548 -0.49235804
#> [43,]  0.46681789 -2.35008723  0.51791349  0.03315946
#> [44,]  0.93175577  0.62763076 -0.30710971  0.74747144
#> [45,]  0.61443622 -0.14710955  0.12406685 -0.05847084
#> [46,]  0.07437105  0.61404255 -2.26632285 -0.06528493
#> [47,] -0.81100212  0.13998822 -1.68623978  1.79358963
#> [48,] -1.27707893 -1.29776514  1.33403923 -0.49725923
#> [49,] -0.46404922 -1.02034982 -1.77632184 -0.95949225
#> [50,]  0.32501631 -1.18450316  0.50678164  0.74430331
#> 
#> $datablocks[[5]]
#>               [,1]        [,2]        [,3]        [,4]
#>  [1,]  1.766147313  0.02962066  0.35531640  0.10324107
#>  [2,] -0.678211162  1.33888730  0.30371010 -0.18159458
#>  [3,]  1.595107019  0.83543250 -1.53531926  0.33042822
#>  [4,] -2.278858833 -1.10111998  0.20999614  0.96736081
#>  [5,] -1.475483909  1.13309592  0.92795679  0.55825461
#>  [6,] -0.960422467 -0.21686738  0.19975507  0.27877252
#>  [7,]  0.303463905  0.90164387 -1.56356663 -0.01639902
#>  [8,]  0.337097437  0.53957586 -1.89085490 -1.22915758
#>  [9,]  0.018406688  0.14777201 -0.85066518 -0.04676166
#> [10,] -0.667154976 -0.84377070  1.40365152  0.37960028
#> [11,]  1.270579331 -2.06888288 -0.03882902  0.64868661
#> [12,] -0.728109662  1.45897269  0.53931803  0.46337210
#> [13,] -0.360554432  0.05650381  0.64604764  0.51388093
#> [14,] -0.261442067 -0.62515399 -1.25723046 -0.39038428
#> [15,]  0.165408125  1.13707795  1.50787032  0.50409706
#> [16,] -0.490931967 -1.19642780 -0.53320273  0.23888444
#> [17,]  0.414236620 -0.35314685 -0.58549313 -0.52523780
#> [18,] -0.666664562 -0.72285266 -0.32736142  0.81336572
#> [19,] -0.808821147  2.92477368  1.55993534 -1.14019671
#> [20,]  0.595683144 -0.97941995  0.39347616 -0.04923284
#> [21,]  0.252508007  0.30546039  0.11999266 -1.58934990
#> [22,] -0.037778383 -1.03458064 -0.28793741 -0.30985155
#> [23,] -1.820147616  0.39151182 -1.69103127 -0.20116597
#> [24,]  1.139789830  0.64969656 -0.50820669  1.62239852
#> [25,] -0.030461759 -1.46446158 -0.93884017 -0.79411118
#> [26,]  1.294764462  0.35988184  1.23173046 -1.02377346
#> [27,]  0.371577896  0.96211327 -0.15080251 -0.87032455
#> [28,] -1.432509821 -0.30771630  2.06747120 -0.43718263
#> [29,]  0.750539015 -0.15244283 -0.52752180 -0.28183224
#> [30,] -0.332157311  0.17570832  0.30430018  0.53629172
#> [31,] -0.858987042  0.98703017  0.56679237  1.25848878
#> [32,] -0.869023085 -0.38656401 -0.25000526 -0.62907456
#> [33,]  0.004512525  0.97136581 -1.30956013  0.51913617
#> [34,] -0.596835100  0.13637401  0.99060119 -1.48835284
#> [35,] -0.584909880 -0.99418664  0.20409091 -0.06969779
#> [36,]  0.708607035  0.29549583 -0.80025094 -0.89144603
#> [37,]  0.862338660  1.06759619  0.90483274  2.79234904
#> [38,] -0.645013760  1.04318020  1.04911802  0.62601155
#> [39,]  0.040744694  0.69930649  1.96365941 -1.55718058
#> [40,] -0.960772233 -1.45891089  0.40509829  3.36985050
#> [41,] -0.150560017 -0.27383113 -0.14347349 -1.09253678
#> [42,]  1.148091012 -0.24604140 -1.39325214 -0.97983959
#> [43,]  2.208246588 -0.92680703 -0.61118470  0.44662621
#> [44,]  1.869644305 -1.37525554 -1.45370139  0.30280700
#> [45,]  1.416818427 -0.25620488  0.88331258 -0.68672065
#> [46,] -0.936067655 -0.46298193 -0.54107194 -1.38419588
#> [47,] -0.945302702  0.25108702 -0.80259543 -0.92993604
#> [48,] -0.367927137 -2.21497045  0.44484051  0.33281296
#> [49,]  1.440071086 -0.21513696  1.45443371  0.25951669
#> [50,] -0.029274437  1.07857026 -0.64534973  0.92930318
#> 
#> 
#> $indices
#> $indices[[1]]
#> [1]  4 13 21 46 50
#> 
#> $indices[[2]]
#> [1]  5  7 18 20 26
#> 
#> $indices[[3]]
#> [1] 10 28 32 33 44
#> 
#> $indices[[4]]
#> [1] 12 17 27 31 34
#> 
#> $indices[[5]]
#> [1]  8 16 29 38 42
#> 
#> $indices[[6]]
#> [1]  1  3  9 23 24
#> 
#> $indices[[7]]
#> [1] 25 30 37 40 41
#> 
#> $indices[[8]]
#> [1]  6 11 14 45 48
#> 
#> $indices[[9]]
#> [1]  2 15 22 35 49
#> 
#> $indices[[10]]
#> [1] 19 36 39 43 47
#> 
#> 
#> $z
#>  [1] 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5
#> 
#> $pars
#> $pars$family
#> [1] "gaussian"
#> 
#> $pars$mu.x
#>  [1]  0.120957242 -0.115567958 -0.199528546  0.036344820 -0.021086819
#>  [6] -0.051201457 -0.165843698  0.047435691 -0.046129358  0.147392903
#> [11] -0.174572801 -0.076242828 -0.198934778  0.052778452  0.196191645
#> [16]  0.004796533  0.274598187  0.143793285 -0.031511097 -0.042023339
#> 
#> $pars$sd.x
#>  [1] 1.1106469 0.9900339 0.9118661 1.0697913 1.0141073 0.9254322 0.9274439
#>  [8] 0.9273745 0.8764985 0.8741732 1.0249524 0.9224177 0.9265121 0.9943654
#> [15] 1.0403696 1.0209507 0.9602860 1.0619451 0.9845285 1.0046962
#> 
#> $pars$mu.y
#> [1] 0.09369818
#> 
#> $pars$sd.y
#> [1] 1.063361
#> 
#> 
#> attr(,"class")
#> [1] "multiridge"

# \donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x %*% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
#>     glmnet multiridge 
#>  0.4692074  0.5982584 
sapply(coef, function(x) mean((beta-x[-1])^2))
#>     glmnet multiridge 
#>  0.1718059  0.1262343 

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
#>     glmnet multiridge 
#>   73.65774   50.10826 
# }