Multi-penalty ridge regression

Fits multi-penalty ridge regression (tuning regularisation hyperparameters and estimating regression coefficients). This is a wrapper function of some functions from the multiridge-package.

Usage

multiridge(
  x,
  y,
  z,
  family = "gaussian",
  foldid = NULL,
  nfolds = 10L,
  penalties = NULL
)

Arguments

x

predictors: \(n \times p\) matrix

y

response: \(n\)-dimensional vector

z

\(p\)-dimensional vector with entries in \(\{1, \ldots, q\}\)

family

character "linear" (or "gaussian"), "logistic" (or "binomial"), or "cox"

foldid

\(n_0\)-dimensional vector containing the fold identifiers

nfolds

integer specifying the number of folds

penalties

\(q\)-dimensional vector of penalty parameters, or NULL (cross-validation)

Value

Returns an object of class "multiridge", a list with the following slots:

slots from IWLSridge() or IWLSCoxridge()

character family with value "gaussian" (also for "linear"), "binomial" (also for "logistic"), or "cox"

\(q\)-dimensional vector penalties containing optimised regularisation hyperparameters (one for each predictor group)

list indices with nfolds slots (one for each cross-validation fold), each containing the indices of the observations

list datablocks with \(q\) slots (one for each predictor group), each containing an \(n_0 \times p_k\) matrix, where \(k \in \{1, \ldots, q\}\)

\(p\)-dimensional group vector z (see argument)

list pars with slots family (see above), the \(p\)-dimensional vectors mu.x and sd.x and the scalars mu.y and sd.y

Details

The number of observations (samples) for training or testing are indicated by \(n_0\) and \(n_1\), respectively, the number of variables (features) is indicated by \(p\), and the number of variable groups is indicated by \(q\). Observations (samples) are indexed by \(i\) in \(\{1, \ldots, n\}\), variables (features) are indexed by \(j\) in \(\{1, \ldots, p\}\), and variable groups are indexed by \(k\) in \(\{1, \ldots, q\}\). The number of variables in the \(k^{\text{th}}\) group is indicated by \(p_k\), with \(\sum_{k=1}^q p_k = p\) for non-overlapping groups.

References

Mark A. van de Wiel, Mirrelijn M. van Nee and Armin Rauschenberger (2021). "Fast cross-validation for multi-penalty high-dimensional ridge regression" Journal of Computational and Graphical Statistics 30(4):835-847. doi:10.1080/10618600.2021.1904962 .

See also

Examples

# minimal example
n <- 50; p <- 20; q <- 5
x <- matrix(rnorm(n * p), nrow = n , ncol = p)
y <- rnorm(n)
z <- rep(seq_len(q), length.out = p)
multiridge(x = x, y = y, z = z)
#> $etas
#>               [,1]
#>  [1,] -0.253154404
#>  [2,]  0.315801992
#>  [3,]  0.011988964
#>  [4,] -0.434316814
#>  [5,]  0.304697568
#>  [6,] -0.039277098
#>  [7,] -0.220986440
#>  [8,]  0.138063727
#>  [9,]  0.230758897
#> [10,]  0.120229843
#> [11,] -0.517314113
#> [12,] -0.083941810
#> [13,] -0.461495797
#> [14,] -0.401382716
#> [15,]  0.164783506
#> [16,] -0.288833968
#> [17,] -0.409590470
#> [18,] -0.907561033
#> [19,] -0.379065558
#> [20,]  1.179565506
#> [21,]  0.127022333
#> [22,] -0.088604688
#> [23,]  0.023666951
#> [24,]  0.350041447
#> [25,] -0.524036446
#> [26,] -0.472395830
#> [27,]  0.472659473
#> [28,] -0.086553761
#> [29,] -0.267082443
#> [30,] -0.215781413
#> [31,] -0.600313492
#> [32,] -0.127416047
#> [33,] -0.164367731
#> [34,] -1.011844753
#> [35,]  0.436619118
#> [36,]  0.383634144
#> [37,] -0.488182668
#> [38,]  0.190865176
#> [39,] -0.400408915
#> [40,]  0.536141827
#> [41,] -0.178695496
#> [42,] -0.140027653
#> [43,]  0.716335363
#> [44,]  0.002221187
#> [45,]  0.729607393
#> [46,]  1.016178659
#> [47,]  0.319193137
#> [48,]  0.823863936
#> [49,]  0.694744327
#> [50,] -0.126052913
#> 
#> $Ypred
#>               [,1]
#>  [1,] -0.253154404
#>  [2,]  0.315801992
#>  [3,]  0.011988964
#>  [4,] -0.434316814
#>  [5,]  0.304697568
#>  [6,] -0.039277098
#>  [7,] -0.220986440
#>  [8,]  0.138063727
#>  [9,]  0.230758897
#> [10,]  0.120229843
#> [11,] -0.517314113
#> [12,] -0.083941810
#> [13,] -0.461495797
#> [14,] -0.401382716
#> [15,]  0.164783506
#> [16,] -0.288833968
#> [17,] -0.409590470
#> [18,] -0.907561033
#> [19,] -0.379065558
#> [20,]  1.179565506
#> [21,]  0.127022333
#> [22,] -0.088604688
#> [23,]  0.023666951
#> [24,]  0.350041447
#> [25,] -0.524036446
#> [26,] -0.472395830
#> [27,]  0.472659473
#> [28,] -0.086553761
#> [29,] -0.267082443
#> [30,] -0.215781413
#> [31,] -0.600313492
#> [32,] -0.127416047
#> [33,] -0.164367731
#> [34,] -1.011844753
#> [35,]  0.436619118
#> [36,]  0.383634144
#> [37,] -0.488182668
#> [38,]  0.190865176
#> [39,] -0.400408915
#> [40,]  0.536141827
#> [41,] -0.178695496
#> [42,] -0.140027653
#> [43,]  0.716335363
#> [44,]  0.002221187
#> [45,]  0.729607393
#> [46,]  1.016178659
#> [47,]  0.319193137
#> [48,]  0.823863936
#> [49,]  0.694744327
#> [50,] -0.126052913
#> 
#> $convergence
#> [1] FALSE
#> 
#> $nIt
#> [1] 1
#> 
#> $Hresl
#> $Hresl$Hmatl
#>               [,1]
#>  [1,] -0.253154404
#>  [2,]  0.315801992
#>  [3,]  0.011988964
#>  [4,] -0.434316814
#>  [5,]  0.304697568
#>  [6,] -0.039277098
#>  [7,] -0.220986440
#>  [8,]  0.138063727
#>  [9,]  0.230758897
#> [10,]  0.120229843
#> [11,] -0.517314113
#> [12,] -0.083941810
#> [13,] -0.461495797
#> [14,] -0.401382716
#> [15,]  0.164783506
#> [16,] -0.288833968
#> [17,] -0.409590470
#> [18,] -0.907561033
#> [19,] -0.379065558
#> [20,]  1.179565506
#> [21,]  0.127022333
#> [22,] -0.088604688
#> [23,]  0.023666951
#> [24,]  0.350041447
#> [25,] -0.524036446
#> [26,] -0.472395830
#> [27,]  0.472659473
#> [28,] -0.086553761
#> [29,] -0.267082443
#> [30,] -0.215781413
#> [31,] -0.600313492
#> [32,] -0.127416047
#> [33,] -0.164367731
#> [34,] -1.011844753
#> [35,]  0.436619118
#> [36,]  0.383634144
#> [37,] -0.488182668
#> [38,]  0.190865176
#> [39,] -0.400408915
#> [40,]  0.536141827
#> [41,] -0.178695496
#> [42,] -0.140027653
#> [43,]  0.716335363
#> [44,]  0.002221187
#> [45,]  0.729607393
#> [46,]  1.016178659
#> [47,]  0.319193137
#> [48,]  0.823863936
#> [49,]  0.694744327
#> [50,] -0.126052913
#> 
#> $Hresl$Mmatl
#>               [,1]
#>  [1,]  0.616343531
#>  [2,]  0.051161705
#>  [3,] -0.203718868
#>  [4,] -0.258296888
#>  [5,] -0.827175500
#>  [6,]  0.321871737
#>  [7,]  0.107706749
#>  [8,] -1.058170491
#>  [9,]  0.935385772
#> [10,]  0.724523826
#> [11,]  0.592992597
#> [12,]  0.203667334
#> [13,]  0.083387458
#> [14,] -0.554651146
#> [15,]  0.861042564
#> [16,] -0.297900554
#> [17,] -0.007796606
#> [18,]  0.631581306
#> [19,]  0.935880701
#> [20,]  0.104701275
#> [21,] -0.592429265
#> [22,]  0.003675660
#> [23,] -0.569111590
#> [24,]  0.243399062
#> [25,] -0.290727089
#> [26,] -0.089736465
#> [27,] -0.403869141
#> [28,] -0.462184544
#> [29,]  0.261108955
#> [30,] -0.275271479
#> [31,] -1.848215103
#> [32,]  0.399574029
#> [33,] -0.889851244
#> [34,] -0.488588752
#> [35,]  0.201764019
#> [36,]  0.965100718
#> [37,] -0.324703319
#> [38,]  1.658506277
#> [39,] -1.963953696
#> [40,] -1.128706117
#> [41,] -0.296652347
#> [42,] -0.665726187
#> [43,]  0.286002953
#> [44,] -1.142743965
#> [45,]  0.779824174
#> [46,]  1.064197163
#> [47,]  1.241144925
#> [48,]  0.019332232
#> [49,]  1.099490347
#> [50,]  0.246813286
#> 
#> 
#> $linearized
#>               [,1]
#>  [1,]  0.363189127
#>  [2,]  0.366963698
#>  [3,] -0.191729904
#>  [4,] -0.692613702
#>  [5,] -0.522477932
#>  [6,]  0.282594638
#>  [7,] -0.113279692
#>  [8,] -0.920106764
#>  [9,]  1.166144669
#> [10,]  0.844753669
#> [11,]  0.075678484
#> [12,]  0.119725524
#> [13,] -0.378108339
#> [14,] -0.956033862
#> [15,]  1.025826070
#> [16,] -0.586734522
#> [17,] -0.417387076
#> [18,] -0.275979727
#> [19,]  0.556815142
#> [20,]  1.284266781
#> [21,] -0.465406932
#> [22,] -0.084929028
#> [23,] -0.545444639
#> [24,]  0.593440509
#> [25,] -0.814763535
#> [26,] -0.562132295
#> [27,]  0.068790331
#> [28,] -0.548738305
#> [29,] -0.005973488
#> [30,] -0.491052893
#> [31,] -2.448528594
#> [32,]  0.272157982
#> [33,] -1.054218975
#> [34,] -1.500433505
#> [35,]  0.638383137
#> [36,]  1.348734862
#> [37,] -0.812885988
#> [38,]  1.849371453
#> [39,] -2.364362611
#> [40,] -0.592564290
#> [41,] -0.475347844
#> [42,] -0.805753841
#> [43,]  1.002338316
#> [44,] -1.140522779
#> [45,]  1.509431567
#> [46,]  2.080375823
#> [47,]  1.560338062
#> [48,]  0.843196168
#> [49,]  1.794234674
#> [50,]  0.120760372
#> 
#> $unpen
#> [1] FALSE
#> 
#> $intercept
#> [1] TRUE
#> 
#> $eta0
#> [1] 0
#> 
#> $X1
#>       [,1]
#>  [1,]    1
#>  [2,]    1
#>  [3,]    1
#>  [4,]    1
#>  [5,]    1
#>  [6,]    1
#>  [7,]    1
#>  [8,]    1
#>  [9,]    1
#> [10,]    1
#> [11,]    1
#> [12,]    1
#> [13,]    1
#> [14,]    1
#> [15,]    1
#> [16,]    1
#> [17,]    1
#> [18,]    1
#> [19,]    1
#> [20,]    1
#> [21,]    1
#> [22,]    1
#> [23,]    1
#> [24,]    1
#> [25,]    1
#> [26,]    1
#> [27,]    1
#> [28,]    1
#> [29,]    1
#> [30,]    1
#> [31,]    1
#> [32,]    1
#> [33,]    1
#> [34,]    1
#> [35,]    1
#> [36,]    1
#> [37,]    1
#> [38,]    1
#> [39,]    1
#> [40,]    1
#> [41,]    1
#> [42,]    1
#> [43,]    1
#> [44,]    1
#> [45,]    1
#> [46,]    1
#> [47,]    1
#> [48,]    1
#> [49,]    1
#> [50,]    1
#> 
#> $family
#> [1] "gaussian"
#> 
#> $penalties
#> [1]    159.43000     22.85752 945834.47404     20.20812    147.02627
#> 
#> $datablocks
#> $datablocks[[1]]
#>              [,1]        [,2]        [,3]        [,4]
#>  [1,] -1.43687688  1.04429075 -1.10324735  0.62722356
#>  [2,]  0.79759324 -1.26065420 -0.06515450  0.21342895
#>  [3,]  2.09725383 -0.47771302 -0.53238576 -1.75667607
#>  [4,]  0.76991739  0.07230418 -1.43159250  0.93264434
#>  [5,] -1.19294300  0.36947504  1.22087082 -0.38148542
#>  [6,]  0.96224072  0.10333471  0.72625350 -0.29850898
#>  [7,] -0.77947014 -0.73597174 -1.53963867  0.74357471
#>  [8,] -0.61716421  0.59436791 -0.52892885  1.10125365
#>  [9,] -2.27610175 -0.52959237 -0.22074489  0.90729279
#> [10,]  1.03501423  0.15237237  2.21281393  0.12648974
#> [11,]  1.32587265  1.68901225 -0.67525961 -1.47470922
#> [12,]  0.45784252  1.09709909  0.27559992  0.25510550
#> [13,]  1.37488056 -0.45778799  0.76512152  0.69339957
#> [14,]  0.78757948  0.08263286 -0.36604558 -0.92621873
#> [15,] -1.53907378  1.02683945 -0.24551963  0.59668733
#> [16,] -1.41619752 -0.75465786 -0.80819509 -1.23777513
#> [17,] -1.18451923  1.71720441 -1.06501049  0.20006375
#> [18,] -1.00337243 -2.93299889 -1.75569070  0.54609639
#> [19,] -0.15555131  0.45447534 -0.34897228 -0.09858656
#> [20,]  0.91201955  1.48268749  0.92289412  0.53626669
#> [21,]  0.28802367  1.19096526 -1.30888579 -0.31817756
#> [22,] -0.23925520  2.24035881  0.61122538  1.62307535
#> [23,]  0.05937920 -0.98416037  2.33122827 -0.24770385
#> [24,]  0.09833636  0.01057199  1.61343521 -2.51683137
#> [25,]  1.56095607  0.12036236 -0.02478160  0.53785591
#> [26,] -1.68980966 -0.74972363 -0.03010261 -0.48647426
#> [27,]  0.74078672 -0.55380752  0.39212398  0.34410491
#> [28,]  0.82078488 -0.72369389 -0.83447721 -1.10369664
#> [29,]  1.58922401  0.27607699 -0.47084485  0.33305970
#> [30,] -0.22776833 -1.11574331  0.25323130 -0.03557288
#> [31,] -0.59517736 -1.04517685  0.71398221  1.81332567
#> [32,] -1.24745257  0.31495170 -0.01879593  0.46499543
#> [33,] -0.65731178 -1.30540208 -0.78875953 -1.58980104
#> [34,]  1.03603798  0.47504906 -0.97388505 -1.19802378
#> [35,]  0.94745518  1.06525725 -0.11529112 -0.92965108
#> [36,] -0.38053191 -0.15536728 -0.24877749  0.88568007
#> [37,]  0.14919126 -0.28530296 -0.01965436 -0.16842570
#> [38,] -0.62077857 -0.08030097  1.08427426 -0.56402545
#> [39,] -0.82253427 -0.49837794 -0.40014072  1.65682386
#> [40,] -0.19030075  0.11486397 -1.18885150 -0.90601194
#> [41,] -0.55272351  0.92923078 -0.04961212 -1.28030780
#> [42,] -1.15613219 -1.28530429  1.43939250  0.30560277
#> [43,]  0.59179770 -1.52497335  1.06478201  0.79427083
#> [44,] -0.47300860  0.67452604  0.09939456 -1.46013708
#> [45,] -0.55483713  0.22170837 -1.44792540  0.63700897
#> [46,] -0.22605688  0.75475673  1.67818630  1.56906000
#> [47,]  1.04783891  0.03646543  0.13467550  0.39657577
#> [48,]  0.63299895  0.15440032 -0.87796694  0.98621625
#> [49,]  0.43626549  0.85109088  0.05161384 -1.58099140
#> [50,]  0.71565842 -1.86002126  1.89403898  0.73260949
#> 
#> $datablocks[[2]]
#>               [,1]        [,2]        [,3]         [,4]
#>  [1,] -0.295844202  0.85114223  0.71129257  0.704774689
#>  [2,]  1.643053550 -1.14318780 -0.75550695 -0.214374060
#>  [3,] -0.721262637  0.66949106  0.99052713  0.027610093
#>  [4,]  0.693039650 -2.78558335 -0.88989366 -2.200223297
#>  [5,]  1.933503289 -0.42511621  1.66527714 -0.233359004
#>  [6,] -0.017054333 -0.39076567 -0.31987427 -0.084346800
#>  [7,] -0.001293668  0.21535162 -0.85634402  0.604728225
#>  [8,] -0.097947716  0.07491790  0.93473090  0.681128757
#>  [9,]  0.625824872 -1.61007474  0.98794408 -0.670160159
#> [10,] -0.824278860 -0.59133445 -1.01081445 -0.152310354
#> [11,]  1.355806005 -0.54749880 -0.08817099 -1.590542987
#> [12,] -0.358312697  1.99094641 -0.06993397 -0.058460289
#> [13,] -0.165246026 -0.64084871  1.48186064 -1.697998778
#> [14,] -0.182118032 -1.41985022  0.30382222 -0.154864685
#> [15,] -0.072759693  1.59168108 -0.29761669  0.207937445
#> [16,]  0.073893791  0.34145804  0.19837775  0.279743116
#> [17,] -2.604993469 -1.08169912 -0.82005353 -0.294196555
#> [18,] -1.043172521  0.09166019  1.23435396 -1.727367926
#> [19,] -1.204562136 -0.26649836 -1.27226382 -0.772333140
#> [20,]  1.273725102  0.51404577  2.04357109  0.410766253
#> [21,] -0.307676059 -0.01422675  0.18520537 -0.023031862
#> [22,] -1.228796946  0.57062930 -0.85385212 -0.355472695
#> [23,] -0.823377395  0.72894471  0.46495278 -0.172012948
#> [24,]  0.670057801 -0.09753552 -0.16074717  0.508970449
#> [25,] -1.671053178  0.99850108 -1.72175191  0.882158115
#> [26,]  0.193131201  2.26618841 -1.03633021 -1.746712435
#> [27,] -1.214997034 -0.39451331  0.02276793  0.669905421
#> [28,]  0.963632095 -0.88008521  1.42332043 -0.055158866
#> [29,] -0.508879323  1.77092497 -0.03127551  1.369641534
#> [30,] -0.293450335  0.94410222 -0.77576465  0.681319081
#> [31,] -1.957613040  0.49258620 -2.20853816  1.340332714
#> [32,]  0.472550560 -0.20804382 -0.24704464 -1.096146355
#> [33,]  0.174160212 -0.44073905 -0.59909041 -0.670337759
#> [34,] -0.285441149  0.46195956  0.61166228 -2.706278658
#> [35,]  0.772279598  0.31810213  0.10655529  1.322411272
#> [36,] -1.229553781  0.47966370 -0.80052349  1.223127115
#> [37,]  0.090595955 -0.54031702 -0.70252910 -0.009115649
#> [38,]  0.953425292 -0.85901390 -0.05162223  0.658333096
#> [39,]  1.297620451 -1.17157780 -0.61399018 -0.797227797
#> [40,]  0.318837587  0.01055284  0.27704417 -0.246429712
#> [41,]  0.755864412 -0.65736191  0.21920620 -0.485244168
#> [42,] -0.548224657  0.10525524 -1.40698818 -0.251465580
#> [43,] -0.756797137  0.04205471  1.87205025  0.974921155
#> [44,]  0.476391441  0.74489129 -0.38241343  0.667858119
#> [45,]  1.089040042 -1.66029140  0.26109846  1.725460829
#> [46,]  1.669949385  0.95632779  0.62396946  1.950192777
#> [47,] -0.711330995 -0.86108841  0.40092717  0.913086266
#> [48,] -0.454608735  0.37987139  2.43356046 -0.662614880
#> [49,]  0.151322525 -0.69882133 -0.03789294  0.816386419
#> [50,]  1.932940938  1.77482302 -1.44325105  0.506994457
#> 
#> $datablocks[[3]]
#>              [,1]        [,2]        [,3]         [,4]
#>  [1,] -1.73502074  0.47153370 -0.47328398 -1.066082208
#>  [2,]  0.24867413 -0.59884448 -0.20062370 -1.762873797
#>  [3,]  0.49455780  0.75257670 -0.18170645 -0.936366660
#>  [4,]  1.29002612 -0.57945177 -0.57626657  0.988581473
#>  [5,]  1.14040008  1.25311474  1.93001764 -1.337104301
#>  [6,]  1.57959018  0.49820287  0.71168830  1.050884902
#>  [7,]  0.53152367 -0.27306699  0.29382739  0.606394996
#>  [8,] -0.61203105  1.46942689 -0.39999748  0.436091963
#>  [9,]  1.19284543  0.89873483  0.98463098 -0.570027492
#> [10,] -1.67927615 -0.35313015  0.23474318  1.120993596
#> [11,] -0.47580233 -0.90712762 -0.73839538  1.796637615
#> [12,] -0.10454780  0.78107458 -0.56734078 -0.400693685
#> [13,]  0.23344982  0.90044406  0.40339047 -1.486711940
#> [14,]  0.70450596 -1.36333559 -1.14239754  0.036038425
#> [15,] -1.28970528 -0.51294506 -0.13003533  0.863820303
#> [16,]  1.05733644 -0.11871238 -0.68212706  0.693173077
#> [17,] -0.26814619 -0.12043250  0.87497673  1.910886219
#> [18,] -0.86785857 -0.40254387  0.43834617  1.800166568
#> [19,] -0.26026426  0.74825589  1.44871982  1.467271330
#> [20,]  0.14182603  1.10780950 -0.81679232  0.100616607
#> [21,] -1.82805429 -0.70989080  0.81316929 -1.121688301
#> [22,]  1.42067065  0.71296517  0.76743878 -0.726641096
#> [23,] -1.05692154 -1.72723604  0.56685702  0.163874685
#> [24,] -1.22375331 -0.73258213  0.05607932  1.085015544
#> [25,]  0.79445801  0.77364896  1.44271730 -0.676072663
#> [26,]  1.14886632 -0.25746555 -1.18890151  0.020636826
#> [27,] -0.05472124 -0.14243295 -1.20457173 -0.370799866
#> [28,] -0.30584870  0.08773644  2.03855176 -0.482613141
#> [29,] -0.77991516 -2.47867938  0.82827233 -1.291655008
#> [30,]  0.40854445  0.27398847  0.17093370 -0.058454559
#> [31,] -0.19711877 -0.22688532 -1.80531710 -0.005290716
#> [32,]  0.57179941  0.58880848  0.39776491 -1.301553210
#> [33,]  0.02020015  0.41151167  1.24095312 -0.627965503
#> [34,] -0.97065036 -1.38303993 -1.23629702 -0.244864541
#> [35,]  0.68298118 -1.47325557 -0.35814978 -1.372469426
#> [36,] -1.87106525 -0.77333578  0.11620846 -0.665249020
#> [37,]  1.23619569 -0.59549168 -0.36611039  0.978038121
#> [38,]  2.35324005  1.32790415  0.51008000 -0.022905874
#> [39,] -1.36045476  0.80704534  0.89922607  0.406589145
#> [40,] -0.62502017  0.35012656 -0.83267774 -0.885195990
#> [41,]  1.26664791 -0.99225275 -0.40903250 -1.360120213
#> [42,] -0.33360386 -0.34645358 -0.53672124  0.644861377
#> [43,] -0.80346266 -2.03526051  2.12658175 -0.985075058
#> [44,]  0.29933345  0.01524847 -0.81535570  0.872998764
#> [45,]  0.23803651  2.78644930 -0.64029954  1.283681762
#> [46,] -0.68464419  1.41542767 -0.54668813 -0.123921536
#> [47,] -1.05718026  0.24268947 -0.47843785  0.036098489
#> [48,]  0.92588565  0.78865590  0.73402474  1.779734760
#> [49,]  0.68005410 -0.82930720 -2.95580925 -0.921987170
#> [50,] -0.21658229  0.46977979 -0.74586314  0.661296422
#> 
#> $datablocks[[4]]
#>              [,1]        [,2]        [,3]        [,4]
#>  [1,]  0.03752097  1.65564852  0.01442710 -1.26757462
#>  [2,] -0.48923156 -0.91849945  0.30937059  1.20066218
#>  [3,]  0.04796406 -0.13681784 -0.17072146 -0.82500545
#>  [4,] -0.60952933 -0.51160352  1.14006972  0.93263332
#>  [5,]  0.88825905  0.12585896 -0.45518931 -1.13957301
#>  [6,] -1.43549768 -0.73019101  0.06371454 -0.31505699
#>  [7,]  2.24121403  0.42928042  1.09678304  0.96478588
#>  [8,]  0.34694971 -0.45388522  1.11886302 -0.62056137
#>  [9,] -0.63444833 -0.40131782 -0.33138018  1.12466375
#> [10,]  1.71366328 -1.40819460 -0.17305564  0.24465621
#> [11,] -1.08846228  1.08478436  0.05626163 -0.37948665
#> [12,]  0.30459086  0.34792351  0.70919283  0.33228625
#> [13,] -0.61536208  0.65471732  0.74456212 -0.95465904
#> [14,]  1.77576691  0.32033652  0.73551039 -1.00223987
#> [15,]  0.37048518  0.68901240 -1.35345272  0.94236717
#> [16,]  0.24211253  1.71707354  0.10217108 -0.10790328
#> [17,]  0.57188749 -2.20570322  1.92714057 -1.44733601
#> [18,] -0.84852085  1.78454377  0.02185343 -0.13171790
#> [19,]  0.43083280 -0.71759948  1.16657403  0.58604806
#> [20,]  0.22430658 -1.13113801 -1.70939515 -0.11307509
#> [21,] -0.18184330 -0.91217658 -1.09587943 -0.66308173
#> [22,] -0.26804336 -0.12882879  0.63494408  0.86530221
#> [23,] -0.08305965  0.22507149 -0.28100808 -0.54434084
#> [24,] -0.40974382  0.15504191 -1.14372476 -0.02960059
#> [25,]  0.32811064 -0.01324818  0.75141568 -0.58664029
#> [26,]  0.38732795 -0.41782940 -0.94063840  0.99022554
#> [27,] -0.69759394 -2.06282917 -2.30264384 -1.60466687
#> [28,] -1.89269564  1.19924701 -0.46411263 -1.26846722
#> [29,]  0.12560920  1.38568830  2.56858609  0.01177563
#> [30,]  1.98594946 -0.04563745 -0.50190022 -1.13533000
#> [31,] -2.26204450  1.71169052  0.87240765  0.92478570
#> [32,]  1.17065571  0.43329285 -1.30304522  0.48405721
#> [33,]  1.13556742  0.01407056 -1.69513356  0.09119270
#> [34,] -1.23221710  2.13529746  0.66874340  0.56482397
#> [35,]  0.76026965 -0.16648739 -0.57823268 -0.16288683
#> [36,] -0.65277379 -0.61216495  0.94136790  1.15316684
#> [37,]  0.46895649 -0.12058657  0.43060814 -2.77432542
#> [38,] -0.19875068 -0.46032912 -0.89428250 -1.30876400
#> [39,]  0.61480647  0.97591458  0.06617088  0.28556310
#> [40,] -0.27664831 -0.91776003 -0.14379937  2.29379483
#> [41,]  0.23660875  1.01828104  0.29284081  0.29035905
#> [42,] -0.33807122 -1.16621865 -0.42486848  0.17990801
#> [43,] -1.89949304  1.54982087 -1.06205623  1.81531532
#> [44,]  0.73700460 -0.91996532  1.14723478 -0.39150082
#> [45,] -0.22022745 -0.40942722  0.84131838  0.85322496
#> [46,] -1.25736479 -0.65759481  0.81642277  0.62854103
#> [47,]  1.97574781 -0.16318057  0.78918877  0.06486019
#> [48,] -0.88574222 -1.29545925 -0.65696700  1.45872111
#> [49,] -0.73378886 -0.42339537 -1.72925374  0.76103858
#> [50,]  0.08898617 -0.10452689 -0.61700283 -1.27096493
#> 
#> $datablocks[[5]]
#>              [,1]        [,2]        [,3]        [,4]
#>  [1,] -0.16672148  0.59122895  1.57908978  0.78322283
#>  [2,]  0.58547073  1.15943582  0.65747464  0.07414766
#>  [3,] -0.68398595 -0.51165265  0.42526102 -0.32234040
#>  [4,]  0.36945541 -1.41458259  1.33888904 -0.06149897
#>  [5,]  0.40655940  2.00802266 -0.85262272 -0.72424881
#>  [6,] -1.79162729  0.48169744  0.69971952  1.01242186
#>  [7,] -0.67612370 -0.74370610 -0.44909975  1.19539178
#>  [8,] -1.51711985  0.26482844 -0.57465171  0.39611949
#>  [9,] -1.63942919  0.47472293 -1.78106716  1.10023364
#> [10,] -0.33083246  0.48133368  1.45453387  0.20505423
#> [11,] -0.46536790 -0.15386574 -0.67416803 -1.08248286
#> [12,]  1.76980467  0.26481413  0.75963369  0.44994435
#> [13,] -1.17212700  0.37073618  0.05765535 -1.53260044
#> [14,] -0.24979097 -0.03826681 -1.60726720  1.08666046
#> [15,]  1.35348503 -0.20964198 -0.49726910  0.14040881
#> [16,]  0.23438581  0.82477768  0.24092043 -0.45703625
#> [17,]  0.25529273 -0.44353906 -1.37967011 -1.20163443
#> [18,] -1.97063468 -0.42224504  0.23528897  0.94035434
#> [19,]  0.69278651  0.18766756 -0.33825057 -0.73554783
#> [20,]  1.09668542  0.35292027  0.29155818 -1.38161155
#> [21,]  0.24642697 -0.70490613 -1.22178595  1.00180123
#> [22,] -0.07393424 -1.62599113  1.22920141 -0.85044569
#> [23,]  1.20196684 -1.06173381  1.23221218 -1.79213474
#> [24,]  1.09835824 -1.56481742  0.23696067 -0.96008162
#> [25,] -0.96201044  3.35416396 -0.59426178  0.95010928
#> [26,] -1.44148249 -0.28831845 -0.01991933  0.74635008
#> [27,]  0.37900407  1.02628873  1.46268490  1.04501320
#> [28,] -0.89010612 -0.20174996  1.90590841  1.11437017
#> [29,]  0.14317065  0.82744083 -0.53004071 -0.20525259
#> [30,]  0.43044453 -1.01799410  1.43858592  1.37424952
#> [31,]  0.94524800  0.14293093  0.04608113  2.26845092
#> [32,] -0.37905194 -0.24227198 -0.65390224 -1.32807957
#> [33,]  0.16387452 -1.34982963 -0.45064905 -0.18377555
#> [34,] -1.76288804 -1.13341956 -0.11201438 -1.03801154
#> [35,] -0.03858937 -0.70848839 -2.22020397  0.23478278
#> [36,]  0.47901325 -0.17592430  0.49496862 -2.03541822
#> [37,]  2.37166419 -0.06853793 -0.34001414 -1.90801622
#> [38,] -0.08909710 -0.26225422 -0.69570325  0.32439076
#> [39,] -0.49558563  0.52499950 -0.42850859  0.10836949
#> [40,] -1.15610308 -0.60381414 -0.40233497 -0.07327428
#> [41,] -1.00598231 -0.68587547 -1.03005920 -1.17409570
#> [42,] -0.22777375 -0.28171666 -0.13164758  0.44104317
#> [43,]  0.22729890 -0.41764830  1.99617406  0.68323785
#> [44,] -0.48979725  1.22327722 -1.45342733  1.11611171
#> [45,] -0.09659198  1.16026698  0.23768113  0.45033907
#> [46,]  1.02296038 -0.80312660  1.06757893  0.53947536
#> [47,]  1.53334001  1.78537671 -0.55768130 -1.06744680
#> [48,]  1.38738931  1.54170197 -0.87178460  0.88745741
#> [49,]  0.10917891 -0.14886658  1.23163937  0.24630751
#> [50,]  1.26948973 -1.76384781 -0.45169654 -0.80078491
#> 
#> 
#> $indices
#> $indices[[1]]
#> [1]  4 13 21 46 50
#> 
#> $indices[[2]]
#> [1]  5  7 18 20 26
#> 
#> $indices[[3]]
#> [1] 10 28 32 33 44
#> 
#> $indices[[4]]
#> [1] 12 17 27 31 34
#> 
#> $indices[[5]]
#> [1]  8 16 29 38 42
#> 
#> $indices[[6]]
#> [1]  1  3  9 23 24
#> 
#> $indices[[7]]
#> [1] 25 30 37 40 41
#> 
#> $indices[[8]]
#> [1]  6 11 14 45 48
#> 
#> $indices[[9]]
#> [1]  2 15 22 35 49
#> 
#> $indices[[10]]
#> [1] 19 36 39 43 47
#> 
#> 
#> $z
#>  [1] 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5 1 2 3 4 5
#> 
#> $pars
#> $pars$family
#> [1] "gaussian"
#> 
#> $pars$mu.x
#>  [1] -0.08352921 -0.04700024 -0.03853104  0.18946867  0.12787907 -0.23652651
#>  [7] -0.13577735 -0.06216008  0.16342921  0.14512218  0.11931943 -0.05699745
#> [13]  0.03146132  0.07330343  0.24562571  0.14369799 -0.32073714  0.08112909
#> [19]  0.05189569 -0.04059173
#> 
#> $pars$sd.x
#>  [1] 0.9553027 1.1570415 0.8248647 1.0527133 0.9813209 1.1642124 1.1454777
#>  [8] 1.0941452 0.9871600 0.9908573 0.9471783 1.0158720 0.9448906 0.9874847
#> [15] 1.2358121 0.8199917 0.9781540 0.9162500 0.9932406 0.9393738
#> 
#> $pars$mu.y
#> [1] -0.08157827
#> 
#> $pars$sd.y
#> [1] 0.9504368
#> 
#> 
#> attr(,"class")
#> [1] "multiridge"

# \donttest{
# simulation
set.seed(1)
n0 <- 100
n1 <- 10000
n <- n0 + n1
p <- c(100, 50)
z <- rep(x = seq_along(p), times = p)
x <- sapply(X = z, FUN = function(x) stats::rnorm(n = n, sd = x))
beta <- stats::rnorm(n = sum(p), mean = 1, sd = 0) *
        stats::rbinom(n = sum(p), size = 1, prob = 0.2)
eta <- x %*% beta
family <- "gaussian"
if (identical(family, "gaussian")) {
  y <- eta + 0.5 * stats::rnorm(n = n, sd = stats::sd(eta))
} else if (identical(family, "binomial")) {
  y <- stats::rbinom(n = n, size = 1, prob = 1 / (1 + exp(-eta)))
} else if (identical(family, "cox")) {
  time <- stats::rexp(n = n, rate = exp(eta))
  status <- stats::rbinom(n = n, prob = 0.5, size = 1)
  y <- survival::Surv(time = time, event = status)
}
cond <- rep(x = c(TRUE, FALSE), times = c(n0, n1))

y_hat <- coef <- list()

# standard ridge regression
object <- glmnet::cv.glmnet(x = x[cond, ], y = y[cond],
                           family = family, alpha = 0)
coef$glmnet <- stats::coef(object = object, s = "lambda.min")
y_hat$glmnet <- stats::predict(object = object, newx = x[!cond, ],
                              type = "response", s = "lambda.min")

# multi-penalty ridge regression
object <- multiridge(x = x[cond, ], y = y[cond], z = z, family = family)
coef$multiridge <- stats::coef(object = object)
y_hat$multiridge <- stats::predict(object = object, newx = x[!cond, ])

# estimation performance
sapply(coef, function(x) stats::cor(beta, x[-1]))
#>     glmnet multiridge 
#>  0.4692074  0.5982584 
sapply(coef, function(x) mean((beta-x[-1])^2))
#>     glmnet multiridge 
#>  0.1718059  0.1262343 

# predictive performance
if (identical(family, "gaussian")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    mean((x-y[!cond])^2))
} else if (identical(family, "binomial")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    pROC::auc(response = y[!cond],
              predictor = as.vector(x),
              levels = c(0, 1),
              direction = "<"))
} else if (identical(family, "cox")) {
  metric <- sapply(X = y_hat, FUN = function(x)
    survival::concordance(y[!cond]~I(-x))$concordance)
}
metric
#>     glmnet multiridge 
#>   73.65774   50.10826 
# }